% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\name{add_to_report,seaMass-method}
\alias{add_to_report,seaMass-method}
\title{Add figure to Rmarkdown report}
\usage{
\S4method{add_to_report}{seaMass}(object, fig, name, title = name, dir = ".")
}
\description{
Add figure to Rmarkdown report
}
