% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{assemble_report,seaMass_sigma-method}
\alias{assemble_report,seaMass_sigma-method}
\title{Generate html report}
\usage{
\S4method{assemble_report}{seaMass_sigma}(object, filepath = "report", title = seaMass::name(object))
}
\description{
Generate html report
}
