% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.hpc.R
\name{bayesprot.hpc}
\alias{bayesprot.hpc}
\title{Add together two numbers.}
\usage{
bayesprot.hpc(dd, nthread, id = "bayesprot.hpc", ...)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
