% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{bayesprot}
\alias{bayesprot}
\title{Add together two numbers.}
\usage{
bayesprot(dd)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
}
