% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{control}
\alias{control}
\title{Control parameters for the BayesProt Bayesian Protein-level quantification model}
\usage{
control(peptide.model = "independent", peptide.prior = "empirical",
  feature.model = "independent", feature.prior = "empirical",
  error.model = "poisson", missingness.model = "censored",
  missingness.threshold = 0, plots = FALSE, model0.npeptide = 3,
  model0.seed = 0, model0.nchain = 1, model0.nwarmup = 256,
  model0.thin = 8, model0.nsample = 128, model.seed = 0,
  model.nchain = 1, model.nwarmup = 256, model.thin = 8,
  model.nsample = 128, nthread = parallel::detectCores(logical =
  FALSE), hpc = NULL)
}
\arguments{
\item{peptide.model}{Either \code{NULL} (no peptide model), \code{single} (single random effect) or \code{independent}
(per-peptide independent random effects; default)}

\item{peptide.prior}{Either \code{NULL} (uninformative) or \code{empirical} (empirical Bayes; default).}

\item{feature.model}{Either \code{single} (single residual) or \code{independent} (per-feature independent residuals; default)}

\item{feature.prior}{Either \code{NULL} (uninformative) or \code{empirical} (empirical Bayes; default)}

\item{error.model}{Either \code{lognormal} or \code{poisson} (default)}

\item{missingness.threshold}{All feature quants below this are treated as missing}

\item{model0.npeptide}{Empirical Bayes model: Proteins with less than this number of peptides are not considered}

\item{model0.seed}{Empirical Bayes model: Random numnber seed}

\item{model0.nchain}{Empirical Bayes model: Number of MCMC chains to run}

\item{model0.nwarmup}{Empirical Bayes model: Number of MCMC warmup iterations to run for each chain}

\item{model0.thin}{Empirical Bayes model: MCMC thinning factor}

\item{model0.nsample}{Empirical Bayes model: Total number of MCMC samples to generate}

\item{model.seed}{Full BayesProt model: Random numnber seed}

\item{model.nchain}{Full BayesProt model: Number of MCMC chains to run}

\item{model.nwarmup}{Full BayesProt model: Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{Full BayesProt model: MCMC thinning factor}

\item{model.nsample}{Full BayesProt model: Total number of MCMC samples to generate}

\item{nthread}{Number of CPU threads to employ}

\item{hpc}{Either \code{NULL} (execute locally), \code{pbs}, \code{sge} or \code{slurm} (submit to HPC cluster)}

\item{control$missingness.model}{Either \code{zero} (NAs set to 0), \code{feature} (NAs set to lowest quant of that feature) or
\code{censored} (NAs modelled as censored between 0 and lowest quant of that feature; default)}

\item{normalisation.model}{Use either \code{NULL} (no normalisation), \code{median} (median) or \code{cov.rob} (robust
covariance estimation)}
}
\value{
\code{control} object to pass to \link{bayesprot}
}
\description{
Control parameters for the BayesProt Bayesian Protein-level quantification model
}
