% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_filter}
\alias{dea_filter}
\title{Add filter to a differential expression list prior to FDR calculation}
\usage{
dea_filter(data, min.peptides = 1, min.peptides.per.condition = 0,
  min.features = 1, min.features.per.condition = 0,
  min.test.samples = 4, min.test.samples.per.condition = 2,
  min.real.samples = 1, min.real.samples.per.condition = 0,
  mixcompdist = "uniform", as.data.table = FALSE)
}
\description{
Add filter to a differential expression list prior to FDR calculation
}
