% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_limma}
\alias{dea_limma}
\title{Univariate differential expression analysis with 'limma::lmFit' linear model and 'limma::eBayes' moderation of standard errors}
\usage{
dea_limma(fit, data = protein_quants(fit), data.design = design(fit),
  condition = "Condition", use.precision = TRUE,
  use.moderation = TRUE, output = "limma", save.intercept = FALSE,
  as.data.table = FALSE, ...)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
