% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_MCMCglmm_pairwise}
\alias{dea_MCMCglmm_pairwise}
\title{Pair-wise mixed-effects univariate differential expression analysis with 'MCMCglmm'}
\usage{
dea_MCMCglmm_pairwise(fit, data.design = design(fit),
  fixed = ~Condition, prior = list(R = list(V = 1, nu = 0.02)), ...)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
