% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_metafor}
\alias{dea_metafor}
\title{Mixed-effects univariate differential expression analysis with 'metafor'}
\usage{
dea_metafor(fit, data.design = design(fit), mods = ~Condition,
  random = ~1 | Sample, ...)
}
\description{
Default is to a one-way ANOVA on the column 'Condition' in 'data.design'.
}
