% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{dea_metafor2}
\alias{dea_metafor2}
\title{Mixed-effects univariate differential expression analysis with 'metafor' (test version)}
\usage{
dea_metafor2(fit, data.design = design(fit), ...)
}
\description{
Mixed-effects univariate differential expression analysis with 'metafor' (test version)
}
