% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_pca}
\alias{dea_pca}
\title{Principal Component Analysis with protein quantification uncertainty}
\usage{
dea_pca(fit, data.design = design(fit), ...)
}
\description{
Principal Component Analysis with protein quantification uncertainty
}
