% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_ttests}
\alias{dea_ttests}
\title{Univariate differential expression analysis using t-tests}
\usage{
dea_ttests(fit, data.design = design(fit), condition = "Condition",
  paired = FALSE, var.equal = TRUE, output = "ttests",
  as.data.table = FALSE, ...)
}
\description{
Student's t-tests are performed pair-wise across the levels of the 'Condition' in 'data.design'. Note this function does not use
BayesProt's standard errors and hence should only be used for comparitative purposes with the other 'dea' methods.
}
