% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{deamass_fit}
\alias{deamass_fit}
\alias{del}
\alias{control}
\alias{design}
\alias{measurements}
\alias{components}
\alias{groups}
\alias{block_refs}
\alias{norm_func}
\alias{dea_func}
\alias{fdr_func}
\alias{squeeze_var_func}
\alias{dist_var_func}
\alias{dist_mean_func}
\alias{summary.deamass_fit}
\alias{timings}
\alias{measurement_vars}
\alias{component_vars}
\alias{assay_vars}
\alias{component_deviations}
\alias{group_quants}
\alias{group_de}
\alias{group_fdr}
\title{Interrogating the \code{deamass} fit object}
\usage{
deamass_fit(dir = "deamass", quiet = FALSE, force = FALSE)

del(fit)

control(fit)

design(fit, as.data.table = F)

measurements(fit, as.data.table = FALSE)

components(fit, as.data.table = FALSE)

groups(fit, as.data.table = FALSE)

block_refs(fit, key = 1)

norm_func(fit, key = 1)

dea_func(fit, key = 1)

fdr_func(fit, key = 1)

squeeze_var_func(fit, key = 1)

dist_var_func(fit, key = 1)

dist_mean_func(fit, key = 1)

\method{summary}{deamass_fit}(fit, groupID, stage = "")

timings(fit, stage = "", as.data.table = F)

measurement_vars(fit, measurementIDs = NULL, summary = TRUE,
  stage = "", blocks = 1:control(fit)$assay.nblock,
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

component_vars(fit, componentIDs = NULL, summary = TRUE, stage = "",
  blocks = 1:control(fit)$assay.nblock,
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

assay_vars(fit, groupIDs = NULL, summary = TRUE, stage = "",
  blocks = 1:control(fit)$assay.nblock,
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

component_deviations(fit, componentIDs = NULL, summary = TRUE,
  stage = "", blocks = 1:control(fit)$assay.nblock,
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

group_quants(fit, groupIDs = NULL, summary = TRUE,
  norm.func.key = ifelse(as.integer(summary) == 0, 1,
  as.integer(summary)), block.refs.key = ifelse(is.null(norm.func.key),
  1, norm.func.key), stage = "", blocks = 1:control(fit)$assay.nblock,
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

group_de(fit, groupIDs = NULL, key = 1, dist.mean.func.key = key,
  norm.func.key = dist.mean.func.key, block.refs.key = norm.func.key,
  as.data.table = FALSE)

group_fdr(fit, groupIDs = NULL, key = 1, dea.func.key = key,
  dist.mean.func.key = dea.func.key,
  norm.func.key = dist.mean.func.key, block.refs.key = norm.func.key,
  as.data.table = FALSE)
}
\arguments{
\item{dir}{directory containing the \link{deamass} fit to read}

\item{fit}{\code{deamass_fit} object created by \code{deamass}.}
}
\description{
Get information from a \link{deamass} fit.
}
