% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_control.R
\docType{class}
\name{delta_control-class}
\alias{delta_control-class}
\alias{delta_control}
\title{Control parameters for seaMass-Δ}
\usage{
delta_control(
  component.deviations = FALSE,
  keep = NULL,
  summarise = "groups",
  plot = c("group.quants.volcano", "group.quants.fdr", "group.quants.de"),
  model = "MCMCglmm",
  nwarmup = 4096,
  thin = 256,
  fdr.model = "ash",
  truth.func = NULL,
  random.seed = 0
)
}
\arguments{
\item{component.deviations}{Set this to \code{TRUE} to do differential expression analysis on the component deviations as weldeltl as the group quants.}

\item{keep}{Outputs to keep MCMC samples for, \code{NULL} or a subset of c("markdown", "group.quants.de", "component.deviations.de")}

\item{summarise}{Outputs to write csv summaries for, \code{NULL} or a subset of \code{c("groups")}}

\item{model}{Either \code{NULL} (no differential expression analysis) or \code{"MCMCglmm"} (MCMCglmm differential expression analysis)}

\item{nwarmup}{Number of MCMC warmup iterations to run for each chain with MCMCglmm differential expression analysis.}

\item{thin}{MCMC thinning factor with MCMCglmm differential expression analysis.}

\item{fdr.model}{Either \code{NULL} (no false discovery rate correction) or \code{"ash"} (ash false discovery rate correction)}

\item{random.seed}{Random number seed}

\item{nthread}{Number of CPU threads to employ}
}
\description{
Define advanced control parameters for the seaMass-Σ Bayesian model.
}
\section{Functions}{
\itemize{
\item \code{delta_control}: Generator function
}}

