% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dinvchi}
\alias{dinvchi}
\title{Probability density function of an inverse scaled chi distribution}
\usage{
dinvchi(x, df, s)
}
\description{
Probability density function of an inverse scaled chi distribution
}
