% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_invchi}
\alias{dist_invchi}
\title{fit scaled inverse chi squared distribution with fitdistrplus, note our input is stdevs not variances!}
\usage{
dist_invchi(value, plots = FALSE, ...)
}
\description{
fit scaled inverse chi squared distribution with fitdistrplus, note our input is stdevs not variances!
}
