% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_invchisq}
\alias{dist_invchisq}
\title{fit scaled inverse chi squared distribution with fitdistrplus, note our values as stdevs not variances!}
\usage{
dist_invchisq(value, plots = FALSE, ...)
}
\description{
fit scaled inverse chi squared distribution with fitdistrplus, note our values as stdevs not variances!
}
