% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_invchisq_fitdistrplus}
\alias{dist_invchisq_fitdistrplus}
\title{fit scaled inverse chi squared distribution with fitdistrplus}
\usage{
dist_invchisq_fitdistrplus(x, plots = FALSE, ...)
}
\description{
fit scaled inverse chi squared distribution with fitdistrplus
}
