% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_lst}
\alias{dist_lst}
\title{fit location-scale t distribution with fitdistrplus using quantile matching}
\usage{
dist_lst(x, plots = FALSE)
}
\description{
fit location-scale t distribution with fitdistrplus using quantile matching
}
