% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_lst_fitdistrplus}
\alias{dist_lst_fitdistrplus}
\title{fit location-scale t distribution with fitdistrplus}
\usage{
dist_lst_fitdistrplus(x, plots = FALSE, ...)
}
\description{
fit location-scale t distribution with fitdistrplus
}
