% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_lst_samples_ash}
\alias{dist_lst_samples_ash}
\title{fit location-scale t distribution for ash (minimum df = 2)}
\usage{
dist_lst_samples_ash(chain, sample, value, ...)
}
\description{
fit location-scale t distribution for ash (minimum df = 2)
}
