% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_normal_fitdistrplus}
\alias{dist_normal_fitdistrplus}
\title{fit normal distribution with fitdistrplus}
\usage{
dist_normal_fitdistrplus(x, plots = FALSE, ...)
}
\description{
fit normal distribution with fitdistrplus
}
