% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_samples_invchi}
\alias{dist_samples_invchi}
\title{fit scaled inverse chi distribution}
\usage{
dist_samples_invchi(chain, sample, value, ...)
}
\description{
Clarke et al 2012 ("A fast robust method for fitting gamma distributions") are right - CvM is about the only robust way to fit gamma distributions!!
}
