% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_samples_robust_normal}
\alias{dist_samples_robust_normal}
\title{fit normal distribution with fitdistrplus}
\usage{
dist_samples_robust_normal(chain, sample, value)
}
\description{
fit normal distribution with fitdistrplus
}
