% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_sf_with_fixed_df1_fitdistrplus}
\alias{dist_sf_with_fixed_df1_fitdistrplus}
\title{fit scaled F distribution with per-datapoint fixed df1 with fitdistrplus}
\usage{
dist_sf_with_fixed_df1_fitdistrplus(x, df1, plots = FALSE, ...)
}
\description{
fit scaled F distribution with per-datapoint fixed df1 with fitdistrplus
}
