% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dlst}
\alias{dlst}
\title{Probability density function of a location-scale t distribution}
\usage{
dlst(x, m, s, df, log = FALSE)
}
\description{
Probability density function of a location-scale t distribution
}
