% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{execute_model}
\alias{execute_model}
\title{execute_model (internal)}
\usage{
execute_model(fit, i, use.priors)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{use.priors}{true or false}

\item{chain}{Number of chain to process.}
}
\description{
execute_model (internal)
}
