% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.plot.R
\name{fdr.plot}
\alias{fdr.plot}
\title{Add together two numbers.}
\usage{
fdr.plot(dds, ylim = 1)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
