% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{fdr}
\alias{fdr}
\title{Filter and FDR-correct a differential expression analysis}
\usage{
fdr(data, min.peptides = 1, min.peptides.per.condition = 0,
  min.features = 1, min.features.per.condition = 0,
  min.test.samples = 4, min.test.samples.per.condition = 2,
  min.real.samples = 1, min.real.samples.per.condition = 0,
  as.data.table = FALSE)
}
\description{
Filter and FDR-correct a differential expression analysis
}
