% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{fdr_ash}
\alias{fdr_ash}
\title{False Discovery Rate correction with ash}
\usage{
fdr_ash(data, mixcompdist = "uniform", use.DF = TRUE,
  as.data.table = FALSE, ...)
}
\description{
False Discovery Rate correction with ash
}
