% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_noncentral_scaled_t_distribution}
\alias{fit_noncentral_scaled_t_distribution}
\title{fit_noncentral_scaled_t_distribution}
\usage{
fit_noncentral_scaled_t_distribution(value)
}
\description{
fit_noncentral_scaled_t_distribution
}
