% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_markdown,seaMass-method}
\alias{generate_markdown,seaMass-method}
\title{Generate Rmarkdown for figure}
\usage{
\S4method{generate_markdown}{seaMass}(object, fig, root, filepath, title = filepath)
}
\description{
Generate Rmarkdown for figure
}
