% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_biolin.R, R/stat_dpm.R, R/stat_lfsr.R
\name{geom_biolin}
\alias{geom_biolin}
\alias{stat_ydpmdensity}
\alias{stat_lfsr}
\title{Biolin plot}
\usage{
geom_biolin(
  mapping = NULL,
  data = NULL,
  stat = "lfsr",
  position = "dodge",
  ...,
  draw_outline = FALSE,
  draw_quantiles = c(0.05, 0.5, 0.95),
  trim = 0.01,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_ydpmdensity(
  mapping = NULL,
  data = NULL,
  geom = "violin",
  position = "dodge",
  ...,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_lfsr(
  mapping = NULL,
  data = NULL,
  geom = "biolin",
  position = "dodge",
  ...,
  trim = TRUE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{draw_quantiles}{If `not(NULL)` (default), draw lines at the given quantiles of the density estimate.}

\item{trim}{If `TRUE` (default), trim the tails of the biolins to the range of the data. If `FALSE`, don't trim the tails.}

\item{geom, stat}{Use to override the default connection between `geom_biolin()` and `stat_ylfdr()`.}

\item{scale}{if "area" (default), all violins have the same area (before trimming
the tails). If "count", areas are scaled proportionally to the number of
observations. If "width", all violins have the same maximum width.}
}
\description{
A biolin plot is a compact display of an empirical local false discovery rate distribution. Based on geom_violin from ggplot2.
}
\section{Computed variables}{

\describe{
  \item{density}{density estimate}
  \item{scaled}{density estimate, scaled to maximum of 1}
  \item{count}{density * number of points - probably useless for violin plots}
  \item{violinwidth}{density scaled for the violin plot, according to area, counts
                     or to a constant maximum width}
  \item{n}{number of points}
  \item{width}{width of violin bounding box}
}
}

