% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.MSstats.R
\name{import.MSstats}
\alias{import.MSstats}
\title{Add together two numbers.}
\usage{
import.MSstats(dd.input)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
