% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.Progenesis.R
\name{import.Progenesis}
\alias{import.Progenesis}
\title{Import Waters Progenesis 'PeptideIon' csv file.}
\usage{
import.Progenesis(file = NULL, only.used = T, data = NULL)
}
\arguments{
\item{file}{Location of the csv file.}

\item{data}{Advanced: Rather than specifying a `file`, you can enter a `data.frame` loaded with `data.table::fread` here.}
}
\value{
A `data.table` formatted for input into `bayesprot`.
}
\description{
`import.Progenesis` reads in a Waters Progenesis 'PeptideIon' csv file for processing with BayesProt
}
