% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ProteomeDiscoverer.R
\name{import.ProteomeDiscoverer}
\alias{import.ProteomeDiscoverer}
\title{Import SCIEX ProteinPilot data.}
\usage{
import.ProteomeDiscoverer(file = NULL, data = NULL, data.runs = NULL)
}
\arguments{
\item{file}{Location of the PeptideSummary.txt file.}

\item{data}{Advanced: Rather than specifying a `file`, you can enter an already loaded `data.frame` here.}

\item{data.runs}{If your study involves more than one input file per run (i.e. because of fractionation) , `data.runs` defines a one-to-many mapping from column 'File' to column 'Run'.}
}
\value{
A `data.table` formatted for input into `bayesprot`.
}
\description{
`import.ProteinPilot` reads in a SCIEX ProteinPilot PeptideSummary.txt file for processing with BayesProt
}
