% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_FragPipe}
\alias{import_FragPipe}
\title{Import data outputed by FragPipe (SILAC only at present)}
\usage{
import_FragPipe(file = NULL, sum.fractions = FALSE, data = NULL)
}
\arguments{
\item{data}{MSstats output
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{seaMass_sigma}.
}
\description{
Reads in a \code{ion_label_quant} dataset processed by FragPipe.
}
