% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_MaxQuant_peptides}
\alias{import_MaxQuant_peptides}
\title{Import MaxQuant LF data}
\usage{
import_MaxQuant_peptides(
  proteinGroups.file = NULL,
  peptides.file = NULL,
  shared = FALSE,
  proteinGroups.data = NULL,
  peptides.data = NULL
)
}
\arguments{
\item{proteinGroups.file}{Location of the \code{proteinGroups.txt} file.}

\item{peptides.file}{Location of the \code{peptides.txt} file.}

\item{shared}{Include shared peptides?}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{bayesprot}.
}
\description{
Reads in MaxQuant \code{peptides.txt} and \code{proteinGroups.txt}for processing with \link{bayesprot}.
}
