% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_MSnbase}
\alias{import_MSnbase}
\title{Import MaxQuant LF data}
\usage{
import_MSnbase(fData, exprs, log2 = TRUE)
}
\arguments{
\item{proteinGroups.file}{Location of the \code{proteinGroups.txt} file.}

\item{evidence.file}{Location of the \code{evidence.txt} file.}

\item{shared}{Include shared peptides?}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{bayesprot}.
}
\description{
Reads in MaxQuant \code{evidence.txt} and \code{proteinGroups.txt}for processing with \link{bayesprot}.
}
