% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_OpenSwath_PyProphet}
\alias{import_OpenSwath_PyProphet}
\title{Import OpenSWATH - PyProphet data}
\usage{
import_OpenSwath_PyProphet(files = NULL, shared = F,
  m_score.cutoff = 0.01, data = NULL)
}
\arguments{
\item{files}{One of more \code{_with_dscore} files to import and merge.}

\item{m_score.cutoff}{Include only features with PyProphet m_score >= than this?}

\item{data}{Advanced: Rather than specifying \code{files}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{bayesprot}.
}
\description{
Reads in a set of \code{_with_dscore} datasets processed by OpenSWATH and PyProphet for processing with \link{bayesprot}.
}
