% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_Progenesis}
\alias{import_Progenesis}
\title{Import Waters Progenesis data}
\usage{
import_Progenesis(file = NULL, used = T, data = NULL)
}
\arguments{
\item{file}{Location of the \code{pep_ion_measurements.csv} file.}

\item{used}{Include only features marked as used by Progenesis?}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{bayesprot}.
}
\description{
Reads in a Waters Progenesis \code{pep_ion_measurements.csv} file for processing with \link{bayesprot}.
}
