% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ProteinPilot}
\alias{import_ProteinPilot}
\title{Import SCIEX ProteinPilot data}
\usage{
import_ProteinPilot(file = NULL, shared = F, min.conf = "auto",
  filter = c("discordant peptide type", "no iTRAQ", "weak signal"),
  data = NULL)
}
\arguments{
\item{file}{Location of the PeptideSummary.txt file.}

\item{shared}{Include shared peptides?}

\item{min.conf}{Features with peptide ID confidence less than \code{min.conf} (between 0 - 100) are filtered out. The default
\code{"auto"} uses the ProteinPilot default threshold.}

\item{filter}{Other filters to use, which can include \code{"discordant peptide type"}, \code{"no iTRAQ" and
\code{"weak signal"}}}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{bayesprot}.
}
\description{
Reads in a SCIEX ProteinPilot \code{PeptideSummary.txt} file for processing with \link{bayesprot}.
}
