% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ProteomeDiscoverer}
\alias{import_ProteomeDiscoverer}
\title{Import Thermo ProteomeDiscoverer data}
\usage{
import_ProteomeDiscoverer(
  file = NULL,
  use.shared.peptides = FALSE,
  use.not.unique = TRUE,
  use.excluded.by.method = TRUE,
  use.redundant = FALSE,
  data = NULL,
  as.data.table = F
)
}
\arguments{
\item{file}{Location of the \code{PSMs.txt} file.}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}

\item{shared}{Include shared peptides?}

\item{used}{Include only measurements marked as used by ProteomeDiscoverer?}
}
\value{
A \link{data.frame} for input into \link{seaMass_sigma}.
}
\description{
Reads in a Thermo ProteomeDiscoverer \code{PSMs.txt} file.
}
