% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importMSstats.R
\name{importMSstats}
\alias{importMSstats}
\title{Add together two numbers.}
\usage{
importMSstats(datafile, fractions)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
}
