% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProgenesis.R
\name{importProgenesis}
\alias{importProgenesis}
\title{Add together two numbers.}
\usage{
importProgenesis(datafile)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
