% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProteinPilot.R
\name{importProteinPilot}
\alias{importProteinPilot}
\title{Add together two numbers.}
\usage{
importProteinPilot(datafile, fractions)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
}
