% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importProteomeDiscoverer.R
\name{importProteomeDiscoverer}
\alias{importProteomeDiscoverer}
\title{Add together two numbers.}
\usage{
importProteomeDiscoverer(datafile)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
}
