% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\name{init_report,seaMass-method}
\alias{init_report,seaMass-method}
\title{Initialise Rmarkdown report}
\usage{
\S4method{init_report}{seaMass}(
  object,
  dir = ".",
  name = seaMass::name(parent(object)),
  title = paste("seaMass report for project", name)
)
}
\description{
Initialise Rmarkdown report
}
