% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_report.R
\name{init_report,seaMass_sigma-method}
\alias{init_report,seaMass_sigma-method}
\title{Initialise Rmarkdown report}
\usage{
\S4method{init_report}{seaMass_sigma}(
  object,
  dir = ".",
  name = seaMass::name(parent(object)),
  title = paste("seaMass report for project", name)
)
}
\description{
Initialise Rmarkdown report
}
