% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{new_assay_design}
\alias{new_assay_design}
\title{Specify study design for imported dataset}
\usage{
new_assay_design(data)
}
\arguments{
\item{data}{\link{data.frame} returned by  \link{import_ProteinPilot}, \link{import_ProteinPilot},
\link{import_ProteomeDiscovery}, \link{import_Progenesis} or \link{import_OpenSWATH}.}
}
\value{
\link{data.frame} that can be edited and passed as parameter \code{data.design} of \link{seaMass_sigma}.
}
\description{
Returns a skeleton \link{data.frame} for customising the study design of an imported dataset.
}
