% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{new_control}
\alias{new_control}
\title{Control parameters for the BayesProt Bayesian model}
\usage{
new_control(peptide.model = "independent", peptide.prior = 1024,
  feature.model = "independent", feature.prior = 3,
  error.model = "poisson", missingness.model = "censored",
  missingness.threshold = 0, model.seed = 0, model.nchain = 4,
  model.nwarmup = 256, model.thin = 1, model.nsample = 1024,
  nthread = parallel::detectCores(logical = FALSE), hpc = NULL)
}
\arguments{
\item{peptide.model}{Either \code{NULL} (no peptide model), \code{single} (single random effect) or \code{independent}
(per-peptide independent random effects; default)}

\item{peptide.prior}{Number of peptides to use for computing Empirical Bayes peptide & feature priors}

\item{feature.model}{Either \code{single} (single residual) or \code{independent} (per-feature independent residuals; default)}

\item{feature.prior}{Minimum number of features per peptide to use for computing Empirical Bayes peptide & feature priors}

\item{error.model}{Either \code{lognormal} or \code{poisson} (default)}

\item{missingness.model}{Either \code{zero} (NAs set to 0), \code{feature} (NAs set to lowest quant of that feature) or
\code{censored} (NAs modelled as censored between 0 and lowest quant of that feature; default)}

\item{missingness.threshold}{All feature quants below this are treated as missing}

\item{model.seed}{Random number seed}

\item{model.nchain}{Number of MCMC chains to run}

\item{model.nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{MCMC thinning factor}

\item{model.nsample}{Total number of MCMC samples to deliver downstream}

\item{nthread}{Number of CPU threads to employ}

\item{hpc}{Either \code{NULL} (execute locally), \code{pbs}, \code{sge} or \code{slurm} (submit to HPC cluster) [TODO]}
}
\value{
\code{bayesprot_control} object to pass to \link{bayesprot}
}
\description{
Control parameters for the BayesProt Bayesian model
}
