% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{new_delta_control}
\alias{new_delta_control}
\title{Control parameters for seaMass-Δ}
\usage{
new_delta_control(
  norm.model = "theta",
  norm.nwarmup = 256,
  norm.thin = 1,
  dea.model = "MCMCglmm",
  dea.nwarmup = 4096,
  dea.thin = 256,
  fdr.model = "ash",
  random.seed = 0,
  nthread = parallel::detectCores()\%/\%2
)
}
\arguments{
\item{norm.model}{Either 'median' normalisation or NULL (none)}

\item{dea.model}{Either 'MCMCglmm' differential expression analysis or NULL (none)}

\item{dea.thin}{MCMC thinning factor for dea_MCMCglmm input}

\item{fdr.model}{Either 'ash' false discovery rate control or NULL (none)}

\item{random.seed}{Random number seed}

\item{nthread}{Number of CPU threads to employ}

\item{model.nchain}{Number of MCMC chains to run}

\item{model.nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{MCMC thinning factor}

\item{model.nsample}{Total number of MCMC samples to deliver downstream}

\item{hpc}{Either \code{NULL} (execute locally), \code{pbs}, \code{sge} or \code{slurm} (submit to HPC cluster) [TODO]}
}
\value{
\code{seaMass_sigma_control} object to pass to \link{sigma}
}
\description{
Define advanced control parameters for the seaMass-Σ Bayesian model.
}
