% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{new_design}
\alias{new_design}
\title{Specify study design for imported dataset}
\usage{
new_design(data)
}
\arguments{
\item{data}{\link{data.frame} returned by \link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}.}
}
\value{
\link{data.frame} that can be edited and passed as parameter \code{data.design} of \link{seamassdelta}.
}
\description{
Returns a skeleton \link{data.frame} for customising the study design of a dataset imported with  \link{import_ProteinPilot} or
\link{import_ProteomeDiscoverer}.
}
