% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_schedule.R
\name{new_hpc_control}
\alias{new_hpc_control}
\title{HPC parameters for executing seaMass Bayesian process on HPC clusters}
\usage{
new_hpc_control(
  queue = NA,
  mem = "64000M",
  wallTime = "12:00:00",
  cpuNum = 14,
  node = 1,
  taskPerNode = 1,
  email = NA
)
}
\arguments{
\item{queue}{Name of the queue on the HPC to submit the jobs to. Different queues are tailored to have different requirements.}

\item{mem}{Amount of memory needed for each task.}

\item{wallTime}{Specify walltime for HPC submission system.}

\item{node}{Number of nodes to use per task.}

\item{taskPerNode}{Number of Nodes to use per task.}

\item{email}{email address to use for notigation of completed jobs on HPC system.}

\item{compress}{Compress HPC submission files into a ZIP file, default value = TRUE.}
}
\value{
hpc.schedule object to pass to \link{seaMass}
}
\description{
Each stage of seamasdelta is split into seperate tasks, currently each task only needs 1 node.
}
