% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{new_hpc_schedule}
\alias{new_hpc_schedule}
\title{HPC parameters for executing seaMass Bayesian model on HPC clusters}
\usage{
new_hpc_schedule(que = "cpu", mem = "6000M", node = 1,
  taskPerNode = 1, output = "NULL", wallTime = "NULL",
  compress = TRUE, email = "UserName@email.com")
}
\arguments{
\item{que}{Name of the que on the HPC to submit the jobs to. Different ques are tailered and have different requirements.}

\item{mem}{Amount of memory needed for each task.}

\item{node}{Number of nodes to use per task.}

\item{taskPerNode}{Number of Nodes to use per task.}

\item{output}{Output path is used to overide fit directory.}

\item{wallTime}{Specify walltime for HPC submission system.}

\item{compress}{Compress HPC submission files into a ZIP file, default value = TRUE.}

\item{email}{email address to use for notigation of completed jobs on HPC system.}
}
\value{
hpc.schedule object to pass to \link{seaMass}
}
\description{
Each stage of seamasdelta is split into seperate tasks, currently each task only needs 1 node.
}
