% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{new_hpcschedule}
\alias{new_hpcschedule}
\title{HPC parameters for executing seaMass Bayesian model on HPC clusters}
\usage{
new_hpcschedule(taskCPU = NULL, que = "cpu", mem = "6G", node = 1,
  taskPerNode = 1, output = "NULL", email = "UserName@email.com")
}
\arguments{
\item{taskCPU}{Number of CPUs to use per job submission. .i.e. equivilent to the number of threads per task.}

\item{que}{Name of the que on the HPC to submit the jobs to. Different ques are tailered and have different requirements.}

\item{mem}{Amount of memory needed for each task.}

\item{node}{Number of nodes to use per task.}

\item{taskPerNode}{Number of Nodes to use per task.}

\item{output}{Output path is used to overide fit directory}

\item{email}{email address to use for notigation of completed jobs on HPC system}
}
\value{
hpc.schedule object to pass to \link{seaMass}
}
\description{
Each stage of seamasdelta is split into seperate tasks, currently each task only needs 1 node.
}
