% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{new_sigma_control}
\alias{new_sigma_control}
\title{Control parameters for seaMass-Σ}
\usage{
new_sigma_control(
  measurement.model = "independent",
  measurement.eb.min = 2,
  component.model = "independent",
  component.eb.min = 3,
  assay.model = "component",
  assay.eb.min = 3,
  assay.eb.nsample = 16,
  error.model = "poisson",
  missingness.model = "censored",
  missingness.threshold = 0,
  model.nchain = 4,
  model.nwarmup = 256,
  model.thin = 1,
  model.nsample = 1024,
  eb.max = 1024,
  random.seed = 0,
  nthread = parallel::detectCores()\%/\%2,
  hpc = NULL
)
}
\arguments{
\item{measurement.model}{Either \code{single} (single residual) or \code{independent} (per-measurement independent residuals; default)}

\item{measurement.eb.min}{Minimum number of measurements per component to use for computing Empirical Bayes priors}

\item{component.model}{Either \code{NULL} (no component model; default), \code{single} (single random effect) or \code{independent}
(per-component independent random effects)}

\item{component.eb.min}{Minimum number of components per group to use for computing Empirical Bayes priors}

\item{assay.model}{Either \code{NULL} (no assay model), \code{measurement} (per-assay independent random effects across measurements)
or \code{component} (per-assay independent random effects across components; default)}

\item{assay.eb.min}{Minimum number of assays per group group to use for computing Empirical Bayes priors}

\item{error.model}{Either \code{lognormal} or \code{poisson} (default)}

\item{missingness.model}{Either \code{zero} (NAs set to 0), \code{measurement} (NAs set to lowest quant of that measurement) or
\code{censored} (NAs modelled as censored between 0 and lowest quant of that measurement; default)}

\item{missingness.threshold}{All measurement quants equal to or below this are treated as missing (default = 0)}

\item{model.nchain}{Number of MCMC chains to run}

\item{model.nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{MCMC thinning factor}

\item{model.nsample}{Total number of MCMC samples to deliver downstream}

\item{random.seed}{Random number seed}

\item{nthread}{Number of CPU threads to employ}

\item{hpc}{Either \code{NULL} (execute locally), \code{pbs}, \code{sge} or \code{slurm} (submit to HPC cluster) [TODO]}
}
\value{
\code{seaMass_sigma_control} object to pass to the \code{control} parameters of \link{seaMass_sigma}
}
\description{
Define advanced control parameters for the seaMass-Σ Bayesian model.
}
