% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_lmRob}
\alias{norm_lmRob}
\title{lmRob normalisation}
\usage{
norm_lmRob(
  fit,
  input = "standardised",
  output = "normalised",
  norm.groups = ".*",
  as.data.table = FALSE
)
}
\description{
Normalisation through robust linear regression
}
