% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_MCMCglmm}
\alias{norm_MCMCglmm}
\title{MCMCglmm normalisation}
\usage{
norm_MCMCglmm(
  fit,
  input = "standardised",
  output = "normalised",
  norm.groups = NULL,
  as.data.table = FALSE
)
}
\description{
Normalisation through a Bayesian linear regression with per-group variance components.
}
