% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_median}
\alias{norm_median}
\title{Median normalisation}
\usage{
norm_median(fit, assay.dilutions = rep_len(1,
  nlevels(design(fit)$Assay)),
  ref.assays = design(fit)$Assay[design(fit)$ref == T],
  ref.proteins = levels(proteins(fit)$Protein), as.data.table = FALSE,
  ...)
}
\description{
blah
}
