% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_quantile}
\alias{norm_quantile}
\title{Quantile normalisation}
\usage{
norm_quantile(fit, input = "standardised", output = "normalised", ...)
}
\description{
A Bayesian version of quantile normalisation.
}
