% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_robust_median}
\alias{norm_robust_median}
\title{Robust Median normalisation}
\usage{
norm_robust_median(fit, data, ref.groups = levels(groups(fit)$Group),
  as.data.table = FALSE)
}
\description{
blah
}
