% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_theta}
\alias{norm_theta}
\title{seaMass-Θ normalisation}
\usage{
norm_theta(
  fit,
  input = "standardised",
  output = "normalised",
  norm.groups = ".*",
  ...
)
}
\description{
Normalisation through a Bayesian linear regression with per-group variance components.
}
