% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{pinaka}
\alias{pinaka}
\title{Cumulative distribution function (CDF) of an inverse nakagami distribution}
\usage{
pinaka(x, s, df)
}
\description{
Cumulative distribution function (CDF) of an inverse nakagami distribution
}
