% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{pinvchi}
\alias{pinvchi}
\title{Cumulative distribution function (CDF) of an inverse scaled inverse chi distribution}
\usage{
pinvchi(x, df, s)
}
\description{
Cumulative distribution function (CDF) of an inverse scaled inverse chi distribution
}
