% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdr.R
\name{plot.fdr}
\alias{plot.fdr}
\title{Add together two numbers.}
\usage{
\method{plot}{fdr}(data, ymax = 0.2)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
