% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_assay_exposures,seaMass_delta-method}
\alias{plot_assay_exposures,seaMass_delta-method}
\title{Add together two numbers.}
\usage{
\S4method{plot_assay_exposures}{seaMass_delta}(
  object,
  data = normalised_group_quants(object),
  data.design = assay_design(object)
)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
