% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_assay_exposures}
\alias{plot_assay_exposures}
\title{Add together two numbers.}
\usage{
plot_assay_exposures(fit, data = normalised_group_quants(fit),
  data.design = assay_design(fit))
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
