% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_sigma.R
\name{plot_assay_stdevs,seaMass_sigma-method}
\alias{plot_assay_stdevs,seaMass_sigma-method}
\title{#' @import data.table
#' @export
#' @include generics.R
setMethod("plot_priors", "seaMass_sigma", function(
  object,
  data = list(
    priors(object, as.data.table = T)[is.na(Assay)][, .(Block, Covariate, s, df)],
    priors(object, as.data.table = T)[is.na(Assay)][, .(Block, Covariate, s = s0, df = df0)],
    rbind(
      measurement_stdevs(object, input = "model0", summary = T, as.data.table = T)[, .(Block, Covariate = "Measurements", value = rinaka(length(s), s, df))],
      component_stdevs(object, input = "model0", summary = T, as.data.table = T)[, .(Block, Covariate = "Components", value = rinaka(length(s), s, df))]
    )
  ),
  horizontal = TRUE,
  draw_outline = list(TRUE, TRUE, FALSE),
  draw_quantiles = list(0.5, NULL, NULL),
  trim = c(0.05, 0.95),
  colour = list("blue", "black", NULL),
  fill = list("lightblue", NULL, "grey"),
  alpha = list(0.5, 0.5, 0.5),
  facets = "Covariate",
  value.label = "stdev log2 quant",
  value.limits = limits_dists(data, trim, c(0, 1)),
  value.length = 160,
  variable.sort.cols = NULL,
  variable.label.cols = "Block",
  variable.interval = 5,
  show.legend = TRUE,
  file = NULL
) {
  return(plot_dists(object, data, horizontal, draw_outline, draw_quantiles, trim, colour, fill, alpha, facets, value.label, value.limits, value.length, variable.sort.cols, variable.label.cols, variable.interval, show.legend, file))
})}
\usage{
\S4method{plot_assay_stdevs}{seaMass_sigma}(
  object,
  assays = NULL,
  data = list(assay_stdevs(object, assays, as.data.table = T)[, list(Block, Assay, s,
    df)], assay_stdevs(object, assays, as.data.table = T)[, list(Block, Assay, s = B.sC,
    df = B.dfC)], assay_stdevs(object, assays, as.data.table = T)[, list(Block, Assay, s
    = B.sM, df = B.dfM)]),
  draw_quantiles = list(0.5, NULL, NULL),
  trim = c(0.05, 0.95),
  colour = list("A.qM", NULL, NULL),
  fill = list(NULL, "darkgreen", "black"),
  alpha = list(0.75, 0.2, 0.2),
  value.label = "stdev log2 quant",
  value.limits = limits_dists(data, trim, include.zero = T, non.negative = T),
  variable.summary.cols = c("Block", "Run", "Channel", "Assay", "RefWeight", "Sample",
    "Condition", "A.qG", "A.qC", "A.qM", "A.qD"),
  variable.label.cols = c("Sample", "Assay", "Block"),
  ...
)
}
\description{
#' @import data.table
#' @export
#' @include generics.R
setMethod("plot_priors", "seaMass_sigma", function(
  object,
  data = list(
    priors(object, as.data.table = T)[is.na(Assay)][, .(Block, Covariate, s, df)],
    priors(object, as.data.table = T)[is.na(Assay)][, .(Block, Covariate, s = s0, df = df0)],
    rbind(
      measurement_stdevs(object, input = "model0", summary = T, as.data.table = T)[, .(Block, Covariate = "Measurements", value = rinaka(length(s), s, df))],
      component_stdevs(object, input = "model0", summary = T, as.data.table = T)[, .(Block, Covariate = "Components", value = rinaka(length(s), s, df))]
    )
  ),
  horizontal = TRUE,
  draw_outline = list(TRUE, TRUE, FALSE),
  draw_quantiles = list(0.5, NULL, NULL),
  trim = c(0.05, 0.95),
  colour = list("blue", "black", NULL),
  fill = list("lightblue", NULL, "grey"),
  alpha = list(0.5, 0.5, 0.5),
  facets = "Covariate",
  value.label = "stdev log2 quant",
  value.limits = limits_dists(data, trim, c(0, 1)),
  value.length = 160,
  variable.sort.cols = NULL,
  variable.label.cols = "Block",
  variable.interval = 5,
  show.legend = TRUE,
  file = NULL
) {
  return(plot_dists(object, data, horizontal, draw_outline, draw_quantiles, trim, colour, fill, alpha, facets, value.label, value.limits, value.length, variable.sort.cols, variable.label.cols, variable.interval, show.legend, file))
})
}
