% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_exposures}
\alias{plot_exposures}
\title{Add together two numbers.}
\usage{
plot_exposures(
  fit,
  data = group_quants(fit, summary = F),
  data.design = design(fit)
)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
