% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plots.R
\name{plot_fdr,seaMass_delta-method}
\alias{plot_fdr,seaMass_delta-method}
\title{Add together two numbers.}
\usage{
\S4method{plot_fdr}{seaMass_delta}(
  object,
  y.max = NULL,
  y.col = "qvalue",
  data.fdr = group_quants_fdr(object),
  output = "ggplot"
)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
