% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fdr}
\alias{plot_fdr}
\title{Add together two numbers.}
\usage{
plot_fdr(data, ymax = 0.2)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
