% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_plot.R
\name{plot_group_quants,seaMass-method}
\alias{plot_group_quants,seaMass-method}
\title{Plot quants for a group}
\usage{
\S4method{plot_group_quants}{seaMass}(object, group)
}
\arguments{
\item{object}{.}

\item{data}{.}

\item{data.summary}{.}

\item{data.design}{.}

\item{contours}{.}

\item{robust}{.}
}
\value{
A ggplot2 object .
}
\description{
Plot quants for a group
}
