% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_pca,seaMass_delta-method}
\alias{plot_pca,seaMass_delta-method}
\title{Robust PCA plot}
\usage{
\S4method{plot_pca}{seaMass_delta}(
  object,
  data = normalised_group_quants(object),
  data.summary = normalised_group_quants(object, summary = T),
  data.design = assay_design(object),
  contours = 1:2,
  robust = TRUE
)
}
\arguments{
\item{object}{.}

\item{data}{.}

\item{data.summary}{.}

\item{data.design}{.}

\item{contours}{.}

\item{robust}{.}
}
\value{
A ggplot2 object.
}
\description{
Robust PCA plot
}
