% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_plot.R
\name{plot_pca_contours,seaMass-method}
\alias{plot_pca_contours,seaMass-method}
\title{Robust PCA plot with uncertainty contours}
\usage{
\S4method{plot_pca_contours}{seaMass}(
  object,
  data.design = assay_design(object),
  variables = NULL,
  input = "model1",
  type = "normalised.group.quants",
  robust = TRUE,
  contours = 1:2,
  aspect.ratio = 3/4,
  labels = 25,
  colour = "Condition",
  fill = "Condition",
  shape = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{object}{.}

\item{data.design}{.}

\item{robust}{.}

\item{contours}{.}

\item{data}{.}

\item{data.summary}{.}
}
\value{
A ggplot2 object .
}
\description{
Robust PCA plot with uncertainty contours
}
