% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plots.R
\name{plot_pr,seaMass_delta-method}
\alias{plot_pr,seaMass_delta-method}
\title{Precision-Recall plot}
\usage{
\S4method{plot_pr}{seaMass_delta}(
  object,
  truth.func,
  plot.fdr = TRUE,
  y.max = NULL,
  legend.nrow = 1,
  y.col = "qvalue",
  width = 1024,
  height = 768,
  data.fdr = group_quants_fdr(object),
  output = "plotly"
)
}
\arguments{
\item{y.max}{.}

\item{data.fdr}{.}
}
\value{
A ggplot2 object.
}
\description{
Precision-Recall plot
}
