% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_priors}
\alias{plot_priors}
\title{Add together two numbers.}
\usage{
plot_priors(DT.vars, DT.priors, ci, xlab = "v", trans = identity,
  inv.trans = identity, show.input = T, random = T)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
