% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\name{plot_robust_pca,seaMass-method}
\alias{plot_robust_pca,seaMass-method}
\title{Plot robust PCA fit}
\usage{
\S4method{plot_robust_pca}{seaMass}(
  object,
  pcs = c(1, 2),
  colour = "A.sC",
  fill = "A.sC",
  shape = "Condition",
  ellipse.probs = c(0.5, 0.95),
  ellipse.alpha = 0.2,
  ellipse.size = 0.2,
  point.size = 2,
  width = 900,
  height = 700,
  data = NULL,
  ...
)
}
\arguments{
\item{object}{.}

\item{data.design}{.}

\item{items}{.}

\item{input}{.}

\item{type}{.}

\item{scale}{.}

\item{robust}{.}
}
\value{
A ggplot2 object .
}
\description{
Plot robust PCA fit
}
