% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{plst}
\alias{plst}
\title{Cumulative distribution function (CDF) of a location-scale t distribution}
\usage{
plst(q, m, s, df, lower.tail = TRUE, log.p = FALSE)
}
\description{
Cumulative distribution function (CDF) of a location-scale t distribution
}
