% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.fdr.R
\name{plt.fdr}
\alias{plt.fdr}
\title{Add together two numbers.}
\usage{
plt.fdr(dds, ylim = 1)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
