% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.pr.R
\name{plt.pr}
\alias{plt.pr}
\title{Add together two numbers.}
\usage{
plt.pr(dd, truth, ymax = NULL)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
