% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.volcano.R
\name{plt.volcano}
\alias{plt.volcano}
\title{Add together two numbers.}
\usage{
plt.volcano(dd, truth = NULL, ci = T)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
