% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.plot.R
\name{pr.plot}
\alias{pr.plot}
\title{Add together two numbers.}
\usage{
pr.plot(dds, truth, ylim = 1)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
