% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.de.R, R/process.de.qprot.R
\name{process.de}
\alias{process.de}
\title{process.de (BayesProt internal function)}
\usage{
process.de()

process.de()
}
\arguments{
\item{datafile}{A number.}
}
\value{
.

The sum of \code{x} and \code{y}.
}
\description{
process.de (BayesProt internal function)

Add together two numbers.
}
