% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.model.R
\name{process.model}
\alias{process.model}
\title{Add together two numbers.}
\usage{
process.model(chain)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
