% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.model1.R
\name{process.model1}
\alias{process.model1}
\title{process.model (BayesProt internal function)}
\usage{
process.model1(chain)
}
\arguments{
\item{chain}{.}
}
\value{
.
}
\description{
process.model (BayesProt internal function)
}
