% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.output.R
\name{process.output}
\alias{process.output}
\title{process.quant (BayesProt internal function)}
\usage{
process.output()
}
\arguments{
\item{input_dir}{.}
}
\value{
.
}
\description{
process.quant (BayesProt internal function)
}
