% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.qprot.R
\name{process.qprot}
\alias{process.qprot}
\title{Add together two numbers.}
\usage{
process.qprot(chain)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
