% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.quant.R
\name{process.quant}
\alias{process.quant}
\title{Add together two numbers.}
\usage{
process.quant(input_dir)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
