% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{process_model}
\alias{process_model}
\title{process_model (internal)}
\usage{
process_model(chain, path.results = ".")
}
\arguments{
\item{chain}{Number of chain to process.}
}
\description{
process_model (internal)
}
