% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_model1}
\alias{process_model1}
\title{process_model1 (internal)}
\usage{
process_model1(fit, chain)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{chain}{Number of chain to process.}
}
\description{
process_model1 (internal)
}
