% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_model2}
\alias{process_model2}
\title{process_model2 (internal)}
\usage{
process_model2(fit, chain)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{chain}{Number of chain to process.}
}
\description{
process_model2 (internal)
}
