% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_plots}
\alias{process_plots}
\title{process_plots (internal)}
\usage{
process_plots(fit, chain)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{chain}{Number of chain to process.}
}
\description{
process_plots (internal)
}
