% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{protein_de}
\alias{protein_de}
\title{Return 'metafor' differential expression as a list of FDR-controlled data tables}
\usage{
protein_de(fit, key = 1, as.data.table = F)
}
\description{
Return 'metafor' differential expression as a list of FDR-controlled data tables
}
