% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{protein_de_ttest}
\alias{protein_de_ttest}
\title{Univariate differential expression analysis with Student's t-test, returning FDR-controlled data table}
\usage{
protein_de_ttest(fit, data.design = design(fit), as.data.table = F)
}
\description{
Univariate differential expression analysis with Student's t-test, returning FDR-controlled data table
}
