% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{qinaka}
\alias{qinaka}
\title{Quantile function (inverse CDF) of an inverse nakagami distribution}
\usage{
qinaka(p, s, df)
}
\description{
Quantile function (inverse CDF) of an inverse nakagami distribution
}
