% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{qinvchi}
\alias{qinvchi}
\title{Quantile function (inverse CDF) of a scaled inverse chi distribution}
\usage{
qinvchi(p, df, s)
}
\description{
Quantile function (inverse CDF) of a scaled inverse chi distribution
}
