% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{qlst}
\alias{qlst}
\title{Quantile function (inverse CDF) of a location-scale t distribution}
\usage{
qlst(p, m, s, df, lower.tail = TRUE, log.p = FALSE)
}
\description{
Quantile function (inverse CDF) of a location-scale t distribution
}
