% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_report.R
\name{render_report,seaMass_sigma-method}
\alias{render_report,seaMass_sigma-method}
\title{Render Rmarkdown report}
\usage{
\S4method{render_report}{seaMass_sigma}(object, dir = ".")
}
\description{
Render Rmarkdown report
}
