% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\name{robust_pca,seaMass-method}
\alias{robust_pca,seaMass-method}
\title{Robust PCA fit}
\usage{
\S4method{robust_pca}{seaMass}(
  object,
  data.design = assay_design(object),
  input = "model1",
  type = "group.quants",
  items = NULL,
  scale = FALSE,
  robust = TRUE,
  ellipses = TRUE,
  as.data.table = FALSE
)
}
\arguments{
\item{object}{.}

\item{data.design}{.}

\item{input}{.}

\item{type}{.}

\item{items}{.}

\item{scale}{.}

\item{robust}{.}
}
\value{
A ggplot2 object .
}
\description{
Robust PCA fit
}
