% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{runs}
\alias{runs}
\alias{runs<-}
\title{Imported dataset run information}
\usage{
runs(data)

runs(data) <- value
}
\arguments{
\item{data}{\link{data.frame} returned by an \code{import} routine.}

\item{value}{Edited \link{data.frame} created by \code{runs}.}
}
\description{
Get or set run and block information from a \link{data.frame} returned by an \code{import} routine.
Used to manually add run and block information to imported datasets.
}
