% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule_pbs-class}
\alias{schedule_pbs-class}
\alias{schedule_pbs}
\title{Schedule seaMass to run on PBS}
\usage{
schedule_pbs(
  submit.prefix = "",
  q = NULL,
  ppn = NULL,
  mem = NULL,
  walltime = NULL,
  l = NULL,
  M = NULL,
  pre = NULL,
  post = NULL
)
}
\arguments{
\item{q}{.}

\item{ppn}{.}

\item{mem}{.}

\item{walltime}{.}

\item{M}{.}
}
\description{
Schedule seaMass to run on PBS
}
\section{Functions}{
\itemize{
\item \code{schedule_pbs}: Generator function
}}

