% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R, R/sigma_block.R
\docType{class}
\name{seaMass-class}
\alias{seaMass-class}
\alias{normalised_group_quants,sigma_block-method}
\title{seaMass object}
\usage{
\S4method{normalised_group_quants}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)
}
\description{
Methods shared between \link{seaMass_sigma}, \link{sigma_block} and \link{seaMass_delta}
}
\section{Functions}{
\itemize{
\item \code{normalised_group_quants,sigma_block-method}: Get the model normalised group quantifications as a \link{data.frame}.
}}

