% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_delta.R, R/sigma_block.R
\docType{class}
\name{seaMass_delta-class}
\alias{seaMass_delta-class}
\alias{seaMass_delta}
\alias{open_delta}
\alias{run,seaMass_delta-method}
\alias{name,seaMass_delta-method}
\alias{filepath,seaMass_delta-method}
\alias{parent,seaMass_delta-method}
\alias{control,seaMass_delta-method}
\alias{assay_design,seaMass_delta-method}
\alias{groups,seaMass_delta-method}
\alias{components,seaMass_delta-method}
\alias{assay_groups,seaMass_delta-method}
\alias{assay_components,seaMass_delta-method}
\alias{measurements,seaMass_delta-method}
\alias{del,seaMass_delta-method}
\alias{de_standardised_group_deviations,seaMass_delta-method}
\alias{de_component_deviations,seaMass_delta-method}
\alias{fdr_standardised_group_deviations,seaMass_delta-method}
\alias{fdr_component_deviations,seaMass_delta-method}
\alias{add_seaMass_spikein_truth}
\alias{add_Navarro_spikein_truth}
\alias{group_means,sigma_block-method}
\title{seaMass-Δ}
\usage{
seaMass_delta(
  fit,
  data.design = assay_design(fit),
  name = "fit",
  control = delta_control(),
  ...
)

open_delta(fit, name = "fit", quiet = FALSE, force = FALSE)

\S4method{run}{seaMass_delta}(object)

\S4method{name}{seaMass_delta}(object)

\S4method{filepath}{seaMass_delta}(object)

\S4method{parent}{seaMass_delta}(object)

\S4method{control}{seaMass_delta}(object)

\S4method{assay_design}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{groups}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{components}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{assay_groups}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{assay_components}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{measurements}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{del}{seaMass_delta}(object)

\S4method{de_standardised_group_deviations}{seaMass_delta}(
  object,
  groups = NULL,
  summary = FALSE,
  type = "standardised.group.deviations",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{de_component_deviations}{seaMass_delta}(
  object,
  type = "component.deviations",
  as.data.table = FALSE
)

\S4method{fdr_standardised_group_deviations}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "standardised.group.deviations",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{fdr_component_deviations}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "component.deviations",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

add_seaMass_spikein_truth(data.fdr)

add_Navarro_spikein_truth(data.fdr)

\S4method{group_means}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{A \link{seaMass_sigma} object as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on raw group-level
quants output by seaMass-Σ
}
\section{Functions}{
\itemize{
\item \code{seaMass_delta}: Runs seaMass-Σ.

\item \code{open_delta}: Open a complete \code{seaMass_delta} run from the supplied \link{seaMass_sigma} fir object and \code{name}.

\item \code{run,seaMass_delta-method}: Run.

\item \code{name,seaMass_delta-method}: Get name.

\item \code{filepath,seaMass_delta-method}: Get path.

\item \code{parent,seaMass_delta-method}: Get the \link{seaMass_sigma} object.

\item \code{control,seaMass_delta-method}: Get the \link{fit_control}.

\item \code{assay_design,seaMass_delta-method}: Get the study design as a \code{data.frame}.

\item \code{groups,seaMass_delta-method}: Get the groups as a \code{data.frame}.

\item \code{components,seaMass_delta-method}: Get the groups as a \code{data.frame}.

\item \code{assay_groups,seaMass_delta-method}: Get the assay groups as a \code{data.frame}.

\item \code{assay_components,seaMass_delta-method}: Get the assay groups as a \code{data.frame}.

\item \code{measurements,seaMass_delta-method}: Get the groups as a \code{data.frame}.

\item \code{del,seaMass_delta-method}: Delete the \code{seaMass_delta} run from disk.

\item \code{de_standardised_group_deviations,seaMass_delta-method}: Get the model differential expression as a \link{data.frame}.

\item \code{de_component_deviations,seaMass_delta-method}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{fdr_standardised_group_deviations,seaMass_delta-method}: Get the model differential expression as a \link{data.frame}.

\item \code{fdr_component_deviations,seaMass_delta-method}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{add_seaMass_spikein_truth}: Add our spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{add_Navarro_spikein_truth}: Add Navarro spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{group_means,sigma_block-method}: Get the model group means as a \link{data.frame}.
}}

