% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_fit.R
\name{seaMass_delta}
\alias{seaMass_delta}
\alias{open_delta_fit}
\alias{del.seaMass_delta_fit}
\alias{control.seaMass_delta_fit}
\alias{sigma_fits.seaMass_delta_fit}
\alias{assay_design.seaMass_delta_fit}
\alias{groups.seaMass_delta_fit}
\alias{components.seaMass_delta_fit}
\alias{unnormalised_group_quants.seaMass_delta_fit}
\alias{normalised_group_quants.seaMass_delta_fit}
\alias{component_deviations.seaMass_delta_fit}
\alias{group_vars.seaMass_delta_fit}
\alias{group_de.seaMass_delta_fit}
\alias{component_deviations_de.seaMass_delta_fit}
\alias{group_fdr.seaMass_delta_fit}
\alias{component_deviations_fdr.seaMass_delta_fit}
\title{seaMass-Δ}
\usage{
seaMass_delta(
  sigma_fits,
  data.design = assay_design(sigma_fits),
  norm.groups = ".*",
  name = sub("^.*\\\\/(.*)\\\\.seaMass", "\\\\1",
    dirname(as.character(sigma_fits[[1]]))),
  control = delta_control(),
  ...
)

open_delta_fit(sigma_fits, name, quiet = FALSE, force = FALSE)

del.seaMass_delta_fit(fit)

control.seaMass_delta_fit(fit)

sigma_fits.seaMass_delta_fit(fit)

assay_design.seaMass_delta_fit(fit, as.data.table = F)

groups.seaMass_delta_fit(fit, as.data.table = FALSE)

components.seaMass_delta_fit(fit, as.data.table = FALSE)

unnormalised_group_quants.seaMass_delta_fit(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "standardised",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

normalised_group_quants.seaMass_delta_fit(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "normalised",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

component_deviations.seaMass_delta_fit(
  fit,
  components = NULL,
  summary = FALSE,
  input = "standardised.component.deviations",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

group_vars.seaMass_delta_fit(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "normalised.group.vars",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

group_de.seaMass_delta_fit(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "de",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

component_deviations_de.seaMass_delta_fit(
  fit,
  components = NULL,
  summary = FALSE,
  input = "de.component.deviations",
  chains = 1:control(fit)@model.nchain,
  as.data.table = FALSE
)

group_fdr.seaMass_delta_fit(fit, input = "fdr", as.data.table = FALSE)

component_deviations_fdr.seaMass_delta_fit(
  fit,
  input = "fdr.component.deviations",
  as.data.table = FALSE
)
}
\arguments{
\item{sigma_fits}{A list of \link{seaMass_sigma_fit} objects, as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta_fit} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on unnormalised group-level
quants output by seaMass-Σ
}
\section{Functions}{
\itemize{
\item \code{del.seaMass_delta_fit}: Delete the directory backing a previously created \code{seaMass_delta} object.

\item \code{control.seaMass_delta_fit}: Returns the \link{seaMass_delta_control} object from an open \code{seaMass_delta_fit} object.

\item \code{sigma_fits.seaMass_delta_fit}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{assay_design.seaMass_delta_fit}: Returns the study design \code{data.frame} from an open \code{seaMass_delta_fit} object.

\item \code{groups.seaMass_delta_fit}: Returns the group metadata from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{components.seaMass_delta_fit}: Returns the component metadata from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{unnormalised_group_quants.seaMass_delta_fit}: Returns the input unnormalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{normalised_group_quants.seaMass_delta_fit}: Returns the normalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{component_deviations.seaMass_delta_fit}: Returns the input standardised component deviations from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{group_vars.seaMass_delta_fit}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{group_de.seaMass_delta_fit}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{component_deviations_de.seaMass_delta_fit}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{group_fdr.seaMass_delta_fit}: Returns the false discovery rate correct differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{component_deviations_fdr.seaMass_delta_fit}: Returns the false discovery rate correct differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.
}}

