% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\docType{class}
\name{seaMass_group_quants-class}
\alias{seaMass_group_quants-class}
\title{seaMass group quants class}
\description{
\link{seaMass_sigma} and \link{seaMass_theta} both expose group_quants method
}
