% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\docType{class}
\name{seaMass_sigma-class}
\alias{seaMass_sigma-class}
\alias{seaMass_sigma}
\alias{open_seaMass_sigma}
\alias{del,seaMass_sigma-method}
\alias{run,seaMass_sigma-method}
\alias{control,seaMass_sigma-method}
\alias{fits,seaMass_sigma-method}
\alias{assay_design,seaMass_sigma-method}
\title{seaMass-Σ}
\usage{
seaMass_sigma(
  data,
  data.design = new_assay_design(data),
  name = "fits",
  run = TRUE,
  zip = FALSE,
  delta = NULL,
  control = sigma_control()
)

open_seaMass_sigma(path = "fits.seaMass", quiet = FALSE, force = FALSE)

\S4method{del}{seaMass_sigma}(object)

\S4method{run}{seaMass_sigma}(object)

\S4method{control}{seaMass_sigma}(object)

\S4method{fits}{seaMass_sigma}(object)

\S4method{assay_design}{seaMass_sigma}(object, as.data.table = FALSE)
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_MaxQuant}, \link{import_OpenSWATH},
\link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}, .}

\item{data.design}{A \link{data.frame} created by \link{new_assay_design} and then customised, which specifies
assay names and block design.}

\item{name}{Name of folder prefix on disk where all intermediate and output data will be stored.}

\item{run}{Run seaMass-Σ now, or just prepare it for later execution?}

\item{zip}{Zip output directory? \code{run=FALSE} and \code{zip=TRUE} are useful in combination when you want
to run seaMass-Σ remotely such as on a HPC cluster.}

\item{control}{A control object created with \link{sigma_control} specifying control parameters for the model.}
}
\value{
A \code{seaMass_sigma} object, which allows access to each block's \link{sigma_fit} object to access
  various metadata and results.
}
\description{
Fits the seaMass-Σ Bayesian group-level quantification model to imported data.
}
\section{Functions}{
\itemize{
\item \code{seaMass_sigma}: Runs seaMass-Σ.

\item \code{open_seaMass_sigma}: Open a complete \code{seaMass_sigma} run from the supplied \code{path}.

\item \code{del,seaMass_sigma-method}: Delete the \code{seaMass_sigma} run from disk.

\item \code{run,seaMass_sigma-method}: Run.

\item \code{control,seaMass_sigma-method}: Get the \link{sigma_control}.

\item \code{fits,seaMass_sigma-method}: Get the list of \link{sigma_fit} obejcts for the blocks.

\item \code{assay_design,seaMass_sigma-method}: Get the study design for all blocks as a \code{data.frame}.
}}

