% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_sigma.R
\name{seaMass_sigma}
\alias{seaMass_sigma}
\title{Fit the seaMass-Σ Bayesian quantification model}
\usage{
seaMass_sigma(
  data,
  data.design = new_design(data),
  summaries = FALSE,
  plots = FALSE,
  name = "fit",
  control = new_sigma_control()
)
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_GroupPilot} or \link{import_ProteomeDiscoverer}.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
assay names and block design.}

\item{summaries}{Generate all summaries}

\item{plots}{Generate all plots}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"fit"}.}

\item{control}{A control object created with \link{new_sigma_control} specifying control parameters for the model.}
}
\value{
A \code{sigma_fit} object that can be interrogated for various results with \code{group_quants},
  \code{component_deviations}, \code{assay_deviations}, \code{component_stdevs}, \code{measurement_stdevs} and \code{assay_stdevs}.
  \code{del} deletes all associated files on disk.
}
\description{
Fit the seaMass-Σ Bayesian quantification model
}
