% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_theta.R
\docType{class}
\name{seaMass_theta-class}
\alias{seaMass_theta-class}
\alias{seaMass_theta}
\alias{open_theta}
\alias{run,seaMass_theta-method}
\alias{name,seaMass_theta-method}
\alias{filepath,seaMass_theta-method}
\alias{parent,seaMass_theta-method}
\alias{control,seaMass_theta-method}
\alias{assay_design,seaMass_theta-method}
\alias{groups,seaMass_theta-method}
\alias{components,seaMass_theta-method}
\alias{assay_groups,seaMass_theta-method}
\alias{assay_components,seaMass_theta-method}
\alias{measurements,seaMass_theta-method}
\alias{blocks,seaMass_theta-method}
\alias{open_thetas,seaMass_theta-method}
\alias{assay_means,seaMass_theta-method}
\alias{normalised_group_quants,seaMass_theta-method}
\title{seaMass-theta}
\usage{
seaMass_theta(
  fit,
  data.design = assay_design(fit),
  name = "fit",
  control = theta_control(),
  ...
)

open_theta(fit, name = "fit", quiet = FALSE, force = FALSE)

\S4method{run}{seaMass_theta}(object)

\S4method{name}{seaMass_theta}(object)

\S4method{filepath}{seaMass_theta}(object)

\S4method{parent}{seaMass_theta}(object)

\S4method{control}{seaMass_theta}(object)

\S4method{assay_design}{seaMass_theta}(object, as.data.table = FALSE)

\S4method{groups}{seaMass_theta}(object, summary = TRUE, as.data.table = FALSE)

\S4method{components}{seaMass_theta}(object, summary = TRUE, as.data.table = FALSE)

\S4method{assay_groups}{seaMass_theta}(object, summary = TRUE, as.data.table = FALSE)

\S4method{assay_components}{seaMass_theta}(object, summary = TRUE, as.data.table = FALSE)

\S4method{measurements}{seaMass_theta}(object, summary = TRUE, as.data.table = FALSE)

\S4method{blocks}{seaMass_theta}(object)

\S4method{open_thetas}{seaMass_theta}(object, quiet = FALSE, force = FALSE)

\S4method{assay_design}{seaMass_theta}(object, as.data.table = FALSE)

\S4method{assay_means}{seaMass_theta}(
  object,
  assays = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{normalised_group_quants}{seaMass_theta}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{A \link{seaMass_sigma} object as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_theta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_theta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on raw group-level
quants output by seaMass-Σ
}
\section{Functions}{
\itemize{
\item \code{seaMass_theta}: Runs seaMass-Σ.

\item \code{open_theta}: Open a complete \code{seaMass_theta} run from the supplied \link{seaMass_sigma} fit object and \code{name}.

\item \code{run,seaMass_theta-method}: Run.

\item \code{name,seaMass_theta-method}: Get name.

\item \code{filepath,seaMass_theta-method}: Get path.

\item \code{parent,seaMass_theta-method}: Get the \link{seaMass_sigma} object.

\item \code{control,seaMass_theta-method}: Get the \link{fit_control}.

\item \code{assay_design,seaMass_theta-method}: Get the study design as a \code{data.frame}.

\item \code{groups,seaMass_theta-method}: Get the groups as a \code{data.frame}.

\item \code{components,seaMass_theta-method}: Get the groups as a \code{data.frame}.

\item \code{assay_groups,seaMass_theta-method}: Get the assay groups as a \code{data.frame}.

\item \code{assay_components,seaMass_theta-method}: Get the assay groups as a \code{data.frame}.

\item \code{measurements,seaMass_theta-method}: Get the groups as a \code{data.frame}.

\item \code{blocks,seaMass_theta-method}: Get the list of \link{sigma_block} obejcts for the blocks.

\item \code{open_thetas,seaMass_theta-method}: Open the list of \link{seaMass_theta} objects.

\item \code{assay_design,seaMass_theta-method}: Get the study design as a \code{data.frame}.

\item \code{assay_means,seaMass_theta-method}: Get the model assay means as a \link{data.frame}.

\item \code{normalised_group_quants,seaMass_theta-method}: Get the model standardised group deviations as a \link{data.frame}.
}}

