% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_block.R, R/theta_block.R
\docType{class}
\name{sigma_block-class}
\alias{sigma_block-class}
\alias{sigma_block}
\alias{filepath,sigma_block-method}
\alias{name,sigma_block-method}
\alias{root,sigma_block-method}
\alias{container,sigma_block-method}
\alias{control,sigma_block-method}
\alias{blocks,sigma_block-method}
\alias{groups,sigma_block-method}
\alias{measurements,sigma_block-method}
\alias{components,sigma_block-method}
\alias{assay_groups,sigma_block-method}
\alias{assay_components,sigma_block-method}
\alias{assay_design,sigma_block-method}
\alias{priors,sigma_block-method}
\alias{summary,sigma_block-method}
\alias{timings,sigma_block-method}
\alias{assay_stdevs,sigma_block-method}
\alias{assay_deviations,sigma_block-method}
\alias{measurement_means,sigma_block-method}
\alias{measurement_stdevs,sigma_block-method}
\alias{component_means,sigma_block-method}
\alias{component_stdevs,sigma_block-method}
\alias{component_deviations,sigma_block-method}
\alias{group_quants,sigma_block-method}
\alias{blocks,theta_block-method}
\title{seaMass-Σ block}
\usage{
\S4method{filepath}{sigma_block}(object)

\S4method{name}{sigma_block}(object)

\S4method{root}{sigma_block}(object)

\S4method{container}{sigma_block}(object)

\S4method{control}{sigma_block}(object)

\S4method{blocks}{sigma_block}(object)

\S4method{groups}{sigma_block}(object, as.data.table = FALSE)

\S4method{measurements}{sigma_block}(object, as.data.table = FALSE)

\S4method{components}{sigma_block}(object, as.data.table = FALSE)

\S4method{assay_groups}{sigma_block}(object, as.data.table = FALSE)

\S4method{assay_components}{sigma_block}(object, as.data.table = FALSE)

\S4method{assay_design}{sigma_block}(object, as.data.table = FALSE)

\S4method{priors}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{summary}{sigma_block}(object, group, chain, input = "model1")

\S4method{timings}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{assay_stdevs}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{assay_deviations}{sigma_block}(
  object,
  assays = NULL,
  summary = TRUE,
  input = "model0",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{measurement_means}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{measurement_stdevs}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_means}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_stdevs}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_deviations}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_quants}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{blocks}{theta_block}(object)
}
\description{
The results of a seaMass-Σ block to a single block, returned by /link{seaMass_sigma} /code{blocks()} function
}
\section{Methods (by generic)}{
\itemize{
\item \code{filepath(sigma_block)}: Get the path.

\item \code{name(sigma_block)}: Get the block name.

\item \code{root(sigma_block)}: Get the \code{seaMass_sigma} object.

\item \code{container(sigma_block)}: Get the \link{seaMass_sigma} object for this block.

\item \code{control(sigma_block)}: Get the \link{sigma_control} object for this block.

\item \code{blocks(sigma_block)}: Get the list of \link{sigma_block} objects for the blocks.

\item \code{groups(sigma_block)}: Get the group metadata as a \link{data.frame}.

\item \code{measurements(sigma_block)}: Get the measurement metadata as a \link{data.frame}.

\item \code{components(sigma_block)}: Get the component metadata as a \link{data.frame}.

\item \code{assay_groups(sigma_block)}: Get the assay groups as a \code{data.frame}.

\item \code{assay_components(sigma_block)}: Get the assay groups as a \code{data.frame}.

\item \code{assay_design(sigma_block)}: Get the study design for this block as a \link{data.frame}.

\item \code{priors(sigma_block)}: Get the priors if computed.

\item \code{summary(sigma_block)}: Print the model summary for a group.

\item \code{timings(sigma_block)}: Get the model timings as a \link{data.frame}.

\item \code{assay_stdevs(sigma_block)}: Get the model assay stdevs as a \link{data.frame}.

\item \code{assay_deviations(sigma_block)}: Get the model assay deviations as a \link{data.frame}.

\item \code{measurement_means(sigma_block)}: Get the model measurement means as a \link{data.frame}.

\item \code{measurement_stdevs(sigma_block)}: Get the model measurement stdevs as a \link{data.frame}.

\item \code{component_means(sigma_block)}: Get the model component means as a \link{data.frame}.

\item \code{component_stdevs(sigma_block)}: Get the model component stdevs as a \link{data.frame}.

\item \code{component_deviations(sigma_block)}: Gets the model component deviations as a \link{data.frame}.

\item \code{group_quants(sigma_block)}: Get the model group quantifications as a \link{data.frame}.

\item \code{blocks(theta_block)}: Get the list of \link{theta_block} objects for the blocks.

}}
