% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_fit.R
\docType{class}
\name{sigma_fit-class}
\alias{sigma_fit-class}
\alias{sigma_fit}
\alias{control,sigma_fit-method}
\alias{assay_design,sigma_fit-method}
\alias{measurements,sigma_fit-method}
\alias{components,sigma_fit-method}
\alias{groups,sigma_fit-method}
\alias{summary,sigma_fit-method}
\alias{timings,sigma_fit-method}
\alias{measurement_vars,sigma_fit-method}
\alias{component_vars,sigma_fit-method}
\alias{component_deviations,sigma_fit-method}
\alias{assay_deviations,sigma_fit-method}
\alias{unnormalised_group_quants,sigma_fit-method}
\title{seaMass-Σ fit}
\usage{
\S4method{control}{sigma_fit}(object)

\S4method{assay_design}{sigma_fit}(object, as.data.table = FALSE)

\S4method{measurements}{sigma_fit}(object, as.data.table = FALSE)

\S4method{components}{sigma_fit}(object, as.data.table = FALSE)

\S4method{groups}{sigma_fit}(object, as.data.table = FALSE)

\S4method{summary}{sigma_fit}(object, group, input = "model1")

\S4method{timings}{sigma_fit}(object, input = "model1", as.data.table = FALSE)

\S4method{measurement_vars}{sigma_fit}(
  object,
  measurements = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_vars}{sigma_fit}(
  object,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_deviations}{sigma_fit}(
  object,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{assay_deviations}{sigma_fit}(
  object,
  measurements = NULL,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{unnormalised_group_quants}{sigma_fit}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)
}
\description{
The results of a seaMass-Σ fit to a single block, returned by /link{seaMass_sigma} /code{fits()} function
}
\section{Methods (by generic)}{
\itemize{
\item \code{control}: Get the \link{sigma_control} object for this fit.

\item \code{assay_design}: Get the study design for this block as a \link{data.frame}.

\item \code{measurements}: Get the measurement metadata as a \link{data.frame}.

\item \code{components}: Get the component metadata as a \link{data.frame}.

\item \code{groups}: Get the group metadata as a \link{data.frame}.

\item \code{summary}: Print the model summary for a group.

\item \code{timings}: Get the model timings as a \link{data.frame}.

\item \code{measurement_vars}: Get the model measurement variances as a \link{data.frame}.

\item \code{component_vars}: Get the model component variances as a \link{data.frame}.

\item \code{component_deviations}: Gets the model component deviations as a \link{data.frame}.

\item \code{assay_deviations}: Get the model assay deviations as a \link{data.frame}.

\item \code{unnormalised_group_quants}: Get the model unnormalised group quantifications as a \link{data.frame}.
}}

