% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sigma_fits}
\alias{sigma_fits}
\alias{del}
\alias{sigma_control}
\alias{design}
\alias{measurements}
\alias{components}
\alias{groups}
\alias{summary.seaMass_sigma_fit}
\alias{timings}
\alias{measurement_vars}
\alias{component_vars}
\alias{assay_vars}
\alias{assay_deviations}
\alias{unnormalised_group_quants}
\title{Interrogating the \code{seamassdelta} fit object}
\usage{
sigma_fits(dirs, quiet = FALSE, force = FALSE)

del(fits)

sigma_control(fit)

design(fit, as.data.table = F)

measurements(fit, as.data.table = FALSE)

components(fit, as.data.table = FALSE)

groups(fit, as.data.table = FALSE)

\method{summary}{seaMass_sigma_fit}(fit, groupID, prefix = "model1")

timings(fit, prefix = "model1", as.data.table = F)

measurement_vars(
  fit,
  measurementIDs = NULL,
  summary.func = dist_invchisq_mcmc,
  prefix = "model1",
  chains = 1:sigma_control(fit)$model.nchain,
  as.data.table = FALSE
)

component_vars(
  fit,
  componentIDs = NULL,
  summary.func = dist_invchisq_mcmc,
  prefix = "model1",
  chains = 1:sigma_control(fit)$model.nchain,
  as.data.table = FALSE
)

assay_vars(
  fit,
  groupIDs = NULL,
  summary.func = dist_invchisq_mcmc,
  prefix = "model1",
  chains = 1:sigma_control(fit)$model.nchain,
  as.data.table = FALSE
)

assay_deviations(
  fit,
  componentIDs = NULL,
  summary.func = dist_lst_mcmc,
  prefix = "model1",
  chains = 1:sigma_control(fit)$model.nchain,
  as.data.table = FALSE
)

unnormalised_group_quants(
  fit,
  groupIDs = NULL,
  summary.func = dist_lst_mcmc,
  prefix = "model1",
  chains = 1:sigma_control(fit)$model.nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{\code{sigma_fits} object created by \code{seamassdelta}.}

\item{dir}{directory containing the \link{seamassdelta} fit to read}
}
\description{
Get information from a \link{seamassdelta} fit.
}
