% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_model.R
\name{sigma_model}
\alias{sigma_model}
\title{sigma_model (internal)}
\usage{
sigma_model(fit, dir, chain = 1)
}
\arguments{
\item{fit}{seaMass_fit object.}

\item{chain}{Number of chain to process.}

\item{use.priors}{true or false}
}
\description{
sigma_model (internal)
}
