% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_process.R
\name{sigma_process0}
\alias{sigma_process0}
\title{sigma_process0 (internal)}
\usage{
sigma_process0(fit, chain)
}
\arguments{
\item{fit}{seamassdelta fit object.}

\item{i}{.}
}
\description{
sigma_process0 (internal)
}
