% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_process.R
\name{sigma_process1}
\alias{sigma_process1}
\title{sigma_process1 (internal)}
\usage{
sigma_process1(fit, chain)
}
\arguments{
\item{fit}{fit object.}

\item{chain}{Number of chain to process.}
}
\description{
sigma_process1 (internal)
}
