% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{squeeze_stdev}
\alias{squeeze_stdev}
\title{Our squeezeVar function - works for small DF unlike Limma squeezeVar, but slower}
\usage{
squeeze_stdev(s, df, use.deconvolution = TRUE, ...)
}
\description{
Our squeezeVar function - works for small DF unlike Limma squeezeVar, but slower
}
