% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{squeeze_var}
\alias{squeeze_var}
\title{Our squeezeVar function - works for small DF unlike Limma squeezeVar, but slower}
\usage{
squeeze_var(v, df, use.deconvolution = TRUE, ...)
}
\description{
Our squeezeVar function - works for small DF unlike Limma squeezeVar, but slower
}
