% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{standardise_blocks}
\alias{standardise_blocks}
\title{Standardise blocks}
\usage{
standardise_blocks(
  fit,
  output = "standardised",
  data.design = design(fit),
  as.data.table = FALSE
)
}
\description{
Make assays across blocks comparable by standardising using RefWeights as the denominator.
}
