% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{standardise_component_deviations}
\alias{standardise_component_deviations}
\title{Standardise component deviations}
\usage{
standardise_component_deviations(
  fit,
  output = "standardised.component.deviations",
  ...
)
}
\description{
Make component deviations comparable between blocks.
}
