% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{standardise_group_quants}
\alias{standardise_group_quants}
\title{Standardise group quants}
\usage{
standardise_group_quants(fit, output = "standardised",
  data.design = assay_design(fit), ...)
}
\description{
Make unnormalised group quants across blocks comparable by standardising using RefWeights as the denominator.
}
