% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{summarise_normal_robust_mcmc}
\alias{summarise_normal_robust_mcmc}
\title{fit normal distribution with fitdistrplus}
\usage{
summarise_normal_robust_mcmc(chainID, mcmcID, value, plots = FALSE)
}
\description{
fit normal distribution with fitdistrplus
}
