% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{summarise_normal_robust_samples}
\alias{summarise_normal_robust_samples}
\title{fit normal distribution with fitdistrplus}
\usage{
summarise_normal_robust_samples(chain, sample, value)
}
\description{
fit normal distribution with fitdistrplus
}
