% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_block.R
\docType{class}
\name{theta_block-class}
\alias{theta_block-class}
\alias{theta_block}
\alias{filepath,theta_block-method}
\alias{name,theta_block-method}
\alias{parent,theta_block-method}
\alias{root,theta_block-method}
\alias{container,theta_block-method}
\alias{control,theta_block-method}
\alias{assay_design,theta_block-method}
\alias{assay_means,theta_block-method}
\alias{group_standards,theta_block-method}
\alias{group_quants,theta_block-method}
\title{seaMass-theta block}
\usage{
\S4method{filepath}{theta_block}(object)

\S4method{name}{theta_block}(object)

\S4method{parent}{theta_block}(object)

\S4method{root}{theta_block}(object)

\S4method{container}{theta_block}(object)

\S4method{control}{theta_block}(object)

\S4method{assay_design}{theta_block}(object, as.data.table = FALSE)

\S4method{assay_means}{theta_block}(
  object,
  assays = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_standards}{theta_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_quants}{theta_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)
}
\description{
The results, returned by /link{seaMass_theta} /code{blocks()} function
}
\section{Methods (by generic)}{
\itemize{
\item \code{filepath}: Get the path.

\item \code{name}: Get the block name.

\item \code{parent}: Get the \link{seaMass_theta} object for this block.

\item \code{root}: Get the \code{seaMass_sigma} object.

\item \code{container}: Get the \link{seaMass_theta} object for this block.

\item \code{control}: Get the \link{theta_control} object for this block.

\item \code{assay_design}: Get the study design for this block as a \link{data.frame}.

\item \code{assay_means}: Get the model assay means as a \link{data.frame}.

\item \code{group_standards}: Get the model normalised group quantifications as a \link{data.frame}.

\item \code{group_quants}: Get the model normalised group quantifications as a \link{data.frame}.
}}

