% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_control.R
\docType{class}
\name{theta_control-class}
\alias{theta_control-class}
\alias{theta_control}
\title{Control parameters for seaMass-theta}
\usage{
theta_control(
  summarise = c("normalised.assay.means", "normalised.group.quants"),
  keep = NULL,
  plot = c("normalised.assay.means", "normalised.group.quants"),
  model = "theta",
  nchain = 8,
  nwarmup = 128,
  thin = 1,
  random.seed = 0
)
}
\arguments{
\item{summarise}{Outputs to write csv summaries for, \code{NULL} or a subset of
\code{c("normalised.assay.means", "normalised.group.quants")}
Note, you must summarise or keep \code{"standardised.group.quants"} if you want to run seaMass-Δ!}

\item{keep}{Outputs to keep MCMC samples for, \code{NULL} or a subset of
\code{c("normalised.assay.means", "normalised.group.quants")}}

\item{plot}{Outputs to plot, \code{NULL} or a subset of
\code{c("normalised.assay.means", "normalised.group.quants")}}

\item{nchain}{Number of MCMC chains to run}

\item{nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{thin}{MCMC thinning factor}

\item{random.seed}{Random number seed}

\item{schedule}{Either \link{schedule_local} (execute locally), \link{schedule_pbs} or \link{schedule_slurm} (prepare for submission to HPC cluster)}
}
\description{
Define control parameters for the seaMass-theta Bayesian model.
}
\section{Functions}{
\itemize{
\item \code{theta_control}: Generator function
}}

