% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update assays to reflect multiple runs}
\usage{
update(data, data.runs = NULL)
}
\arguments{
\item{data}{Result of an `import.` BayesProt function}

\item{data.runs}{A `data.frame` defining a one-to-many mapping from column 'File' to column 'Run'.}
}
\value{
A `data.frame` formatted for input into `bayesprot` with added run information.
}
\description{
Updates the assay column of imported data to include information on multiple runs. Only necessary for TMT and iTraq studies with more than one multiplex .
}
