% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.hpc.R
\name{bayesprot.hpc}
\alias{bayesprot.hpc}
\title{Bayesian Protein-level Quantification for Proteomics (HPC version)}
\usage{
bayesprot.hpc(dd, id = "bayesprot.hpc", de.design = NULL,
  de.paired = F, ref.assays = levels(dd$Assay), missing = "censored",
  truth = NULL, prior.scale = 1, study.nsample = 1024,
  study.nwarmup = 256, study.thin = 1, study.nchain = 1,
  study.seed = 0, study.npeptide = 5, quant.nsample = 1024,
  quant.nwarmup = 256, quant.thin = 1, quant.nchain = 1,
  quant.seed = 0, qprot = F, qprot.path = "",
  qprot.nsample = 12000, qprot.nwarmup = 2000, qprot.seed = 0,
  nthread = parallel::detectCores(), ...)
}
\arguments{
\item{dd}{dataset returned by a bayesprot::import.() function}

\item{id}{path and filename for the output files (if path omitted, current working directory is used)}

\item{de.design}{.}

\item{de.paired}{.}

\item{ref.assays}{.}

\item{missing}{.}

\item{truth}{.}

\item{prior.scale}{.}

\item{study.nsample}{.}

\item{study.nwarmup}{.}

\item{study.thin}{.}

\item{study.nchain}{.}

\item{study.seed}{.}

\item{study.npeptide}{.}

\item{quant.nsample}{.}

\item{quant.nwarmup}{.}

\item{quant.thin}{.}

\item{quant.nchain}{.}

\item{quant.seed}{.}

\item{qprot}{.}

\item{qprot.path}{.}

\item{qprot.nsample}{.}

\item{qprot.nwarmup}{.}

\item{qprot.seed}{.}

\item{nthread}{.}

\item{...}{any other HPC backend parameters}
}
\value{
Lots of interesting stuff.
}
\description{
Bayesian Protein-level Quantification for Proteomics (HPC version)
}
