% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R, R/fit.R
\name{bayesprot}
\alias{bayesprot}
\alias{protein_quants}
\alias{peptide_deviations}
\alias{peptide_stdevs}
\alias{feature_stdevs}
\alias{de_metafor}
\alias{de_mice}
\alias{del}
\title{Fit the BayesProt Bayesian Protein-level quantification model}
\usage{
bayesprot(data, data.design = design(data),
  normalisation.proteins = levels(data$Protein), plot = F,
  output = "bayesprot", control = bayesprot::control())

protein_quants(fit)

peptide_deviations(fit)

peptide_stdevs(fit)

feature_stdevs(fit)

de_metafor(fit)

de_mice(fit)

del(fit)
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{design} and then customised, which specifies
assay-level study design, including reference assays, assay info and any covariates for optional differential expression
analysis. By default, all assays are set as reference channels, which is appropriate only for label-free studies
and fully-blocked iTraq/TMT/SILAC designs.}

\item{normalisation.proteins}{Proteins to use in the normalisation; default is all proteins.}

\item{plot}{Generate all plots (todo)}

\item{output}{Folder on disk whether all intermediate and output data will be stored; default is \code{"bayesprot"}.}

\item{control}{A \link{control} object specifying control parameters for the model.}

\item{fit}{\code{bayesprot_fit} object created by \code{bayesprot}.}
}
\value{
A \code{bayesprot.fit} object that can be interrogated for various results with \code{protein_quants},
  \code{peptide_deviations}, \code{peptide_stdevs}, \code{feature_stdevs}, \code{de_metafor} and \code{de_mice}. \code{del}
  deletes all associated files on disk.
}
\description{
Fit the BayesProt Bayesian Protein-level quantification model
}
