% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{bayesprot}
\alias{bayesprot}
\title{Bayesian Protein-level modelification for Proteomics}
\usage{
bayesprot(dd, id = "bayesprot", assay.refs = levels(dd$Assay),
  assay.digests = levels(dd$Assay), assay.samples = levels(dd$Assay),
  assay.conditions = NULL, de.paired = F, de.mcmc = F,
  missing = "censored", plots = F, model0.min.npeptide = 2,
  model0.nsample = 1024, model0.nwarmup = 256, model0.thin = 1,
  model0.nchain = 1, model0.seed = 0, model.nsample = 1024,
  model.nwarmup = 256, model.thin = 1, model.nchain = 1,
  model.seed = 0, qprot = F, qprot.path = "",
  qprot.nsample = 10000, qprot.nwarmup = 2000, qprot.seed = 0,
  nthread = parallel::detectCores())
}
\arguments{
\item{dd}{dataset returned by a bayesprot::import.() function}

\item{id}{path and filename for the output files (if path omitted, current working directory is used)}

\item{assay.refs}{= levels(dd$Assay),}

\item{assay.digests}{= levels(dd$Assay),}

\item{assay.samples}{= levels(dd$Assay),}

\item{assay.conditions}{= NULL,}

\item{de.paired}{.}

\item{de.mcmc}{.}

\item{missing}{.}

\item{plots}{.}

\item{model0.min.npeptide}{.}

\item{model0.nsample}{.}

\item{model0.nwarmup}{.}

\item{model0.thin}{.}

\item{model0.nchain}{.}

\item{model0.seed}{.}

\item{model.nsample}{.}

\item{model.nwarmup}{.}

\item{model.thin}{.}

\item{model.nchain}{.}

\item{model.seed}{.}

\item{qprot}{= .}

\item{qprot.path}{.}

\item{qprot.nsample}{.}

\item{qprot.nwarmup}{.}

\item{qprot.seed}{.}

\item{nthread}{.}
}
\value{
Lots of interesting stuff.
}
\description{
Bayesian Protein-level modelification for Proteomics
}
