% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{bayesprot}
\alias{bayesprot}
\title{Fit the BayesProt Bayesian Protein-level quantification model}
\usage{
bayesprot(data, data.design = new_design(data),
  normalisation.proteins = levels(data$Protein), de.func = NULL,
  plots = F, output = "bayesprot", control = new_control())
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{design} and then customised, which specifies
assay-level study design, including reference assays, assay info and any covariates for optional differential expression
analysis. By default, all assays are set as reference channels, which is appropriate only for label-free studies
and fully-blocked iTraq/TMT/SILAC designs.}

\item{normalisation.proteins}{Proteins to use in the normalisation; default is all proteins.}

\item{de.func}{List of differential expression analysis functions to run}

\item{output}{Folder on disk whether all intermediate and output data will be stored; default is \code{"bayesprot"}.}

\item{control}{A control object created with \link{new_control} specifying control parameters for the model.}

\item{plot}{Generate all plots (todo)}
}
\value{
A \code{bayesprot_fit} object that can be interrogated for various results with \code{protein_quants},
  \code{peptide_deviations}, \code{peptide_stdevs}, \code{feature_stdevs}, \code{de_metafor} and \code{de_mice}. \code{del}
  deletes all associated files on disk.
}
\description{
Fit the BayesProt Bayesian Protein-level quantification model
}
