% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesprot.R
\name{bayesprot}
\alias{bayesprot}
\title{Add together two numbers.}
\usage{
bayesprot(dd, id = "input", ref.assays = levels(dd$Assay),
  missing = "censored", ...)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
