% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bayesprot_fit}
\alias{bayesprot_fit}
\alias{del}
\alias{design}
\alias{control}
\alias{features}
\alias{peptides}
\alias{proteins}
\alias{ref_assays}
\alias{exposures}
\alias{protein_quants}
\alias{peptide_deviations}
\alias{peptide_vars}
\alias{feature_vars}
\alias{summary.bayesprot_fit}
\alias{timings}
\title{Interrogating the \code{bayesprot} fit object}
\usage{
bayesprot_fit(dir = "bayesprot", force = F)

del(fit)

design(fit, as.data.table = F)

control(fit)

features(fit, as.data.table = F)

peptides(fit, as.data.table = F)

proteins(fit, as.data.table = F)

ref_assays(fit, key = 1, as.data.table = F)

exposures(fit, ref.assays.key = 1, key = 1, as.data.table = F)

protein_quants(fit, proteinIDs = NULL, ref.assays.key = 1,
  exposures.key = 1, summary = T, as.data.table = F, parallel = F,
  stage = 2)

peptide_deviations(fit, proteinIDs = NULL, summary = T,
  as.data.table = F, stage = 2, parallel = F)

peptide_vars(fit, peptideIDs = NULL, summary = T, as.data.table = F,
  stage = 2, parallel = F)

feature_vars(fit, featureIDs = NULL, summary = T, as.data.table = F,
  stage = 2, parallel = F)

\method{summary}{bayesprot_fit}(fit, protein = NULL, proteinID = NULL)

timings(fit, as.data.table = F)
}
\arguments{
\item{dir}{directory containing the \link{bayesprot} fit to read}

\item{fit}{\code{bayesprot_fit} object created by \code{bayesprot}.}
}
\description{
Get information from a \link{bayesprot} fit.
}
