% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bayesprot_fit}
\alias{bayesprot_fit}
\alias{del}
\alias{design}
\alias{control}
\alias{features}
\alias{peptides}
\alias{proteins}
\alias{exposures}
\alias{protein_quants}
\alias{peptide_deviations}
\alias{peptide_stdevs}
\alias{feature_stdevs}
\alias{timings}
\title{Interrogating the \code{bayesprot} fit object}
\usage{
bayesprot_fit(dir, force = F)

del(fit)

design(fit, as.data.table = F)

control(fit)

features(fit, as.data.table = F)

peptides(fit, as.data.table = F)

proteins(fit, as.data.table = F)

exposures(fit, key = 1, as.data.table = F)

protein_quants(fit, ref.assays = design(fit)$Assay[design(fit)$ref == T],
  proteinIDs = NULL, proteins = NULL, summary = T,
  as.data.table = F)

peptide_deviations(fit, proteinID = NULL, protein = NULL,
  summary = T, as.data.table = F)

peptide_stdevs(fit, proteinID = NULL, protein = NULL, summary = T,
  as.data.table = F)

feature_stdevs(fit, proteinID = NULL, protein = NULL, summary = T,
  as.data.table = F)

timings(fit, as.data.table = F)
}
\arguments{
\item{dir}{directory containing the \link{bayesprot} fit to read}

\item{fit}{\code{bayesprot_fit} object created by \code{bayesprot}.}
}
\description{
Get information from a \link{bayesprot} fit.
}
