% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_limma}
\alias{dea_limma}
\title{Univariate differential expression analysis with 'limma::lmFit' linear model and 'limma::eBayes' moderation of standard errors}
\usage{
dea_limma(fit, data = group_quants(fit), data.design = design(fit),
  condition = "Condition", output = "limma", save.intercept = FALSE,
  as.data.table = FALSE, ..., use.moderation = TRUE,
  use.precision = TRUE)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
 todo: see http://www.metafor-project.org/doku.php/tips:rma_vs_lm_lme_lmer
 todo: https://stats.stackexchange.com/questions/142685/equivalent-to-welchs-t-test-in-gls-framework
}
