% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_MCMCglmm}
\alias{dea_MCMCglmm}
\title{Mixed-effects univariate differential expression analysis with 'MCMCglmm'}
\usage{
dea_MCMCglmm(
  fit,
  input = "normalised",
  output = "de",
  data.design = assay_design(fit),
  type = "group.quants",
  emmeans.args = list(revpairwise ~ Condition),
  fixed = ~Condition,
  random = ~idh(Condition):Sample,
  rcov = ~units,
  start = NULL,
  prior = list(G = list(G1 = list(V = diag(nlevels(factor(data.design$Condition))), nu
    = nlevels(factor(data.design$Condition)), alpha.mu = rep(0,
    nlevels(factor(data.design$Condition))), alpha.V = diag(25^2,
    nlevels(factor(data.design$Condition))))), R = list(V = 1, nu = 2e-04)),
  tune = NULL,
  pedigree = NULL,
  nodes = "ALL",
  scale = TRUE,
  pr = FALSE,
  pl = FALSE,
  DIC = TRUE,
  saveX = TRUE,
  saveZ = TRUE,
  saveXL = TRUE,
  slice = FALSE,
  ginverse = NULL,
  trunc = FALSE,
  ...
)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
