% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_MCMCglmm}
\alias{dea_MCMCglmm}
\title{Mixed-effects univariate differential expression analysis with 'metafor'}
\usage{
dea_MCMCglmm(fit, data.design = design(fit), fixed = ~1, ...)
}
\description{
Mixed-effects univariate differential expression analysis with 'metafor'
}
