% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_MCMCglmm}
\alias{dea_MCMCglmm}
\title{Mixed-effects univariate differential expression analysis with 'MCMCglmm'}
\usage{
dea_MCMCglmm(
  fit,
  data = normalised_group_quants(fit, summary = T),
  data.design = design(fit),
  fixed = ~Condition,
  random = NULL,
  rcov = ~idh(Condition):Sample,
  prior = list(R = list(V = diag(nlevels(factor(data.design$Condition))), nu = 2e-04)),
  em.func = function(...) emmeans::emmeans(..., pairwise ~ Condition),
  output = NULL,
  as.data.table = FALSE,
  dist.mean.func = dist_lst_mcmc,
  ...
)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
