% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_MCMCglmm}
\alias{dea_MCMCglmm}
\title{Mixed-effects univariate differential expression analysis with 'MCMCglmm'}
\usage{
dea_MCMCglmm(fit, data.design = design(fit), condition = "Condition",
  fixed = as.formula(paste("~", condition)), prior = list(R = list(V =
  1, nu = 0.02)), use.SE = TRUE, output = "MCMCglmm",
  save.intercept = FALSE, as.data.table = FALSE, ...)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
