% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{dea_metafor}
\alias{dea_metafor}
\title{Mixed-effects univariate differential expression analysis with 'metafor'}
\usage{
dea_metafor(fit, data.design = design(fit), condition = "Condition",
  mods = as.formula(paste("~", condition)), random = ~1 | Sample,
  output = "metafor", save.intercept = FALSE, as.data.table = FALSE,
  ...)
}
\description{
The model is performed pair-wise across the levels of the 'Condition' in 'data.design'. Default is a standard Student's t-test model.
}
