% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{dea_metafor}
\alias{dea_metafor}
\title{Mixed-effects univariate differential expression analysis with 'metafor'}
\usage{
dea_metafor(fit, data.design = design(fit), ...)
}
\description{
Mixed-effects univariate differential expression analysis with 'metafor'
}
