% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\docType{class}
\name{delta_control-class}
\alias{delta_control-class}
\alias{delta_control}
\title{Control parameters for seaMass-Δ}
\usage{
delta_control(
  component.deviations = FALSE,
  norm.model = "theta",
  norm.nwarmup = 256,
  norm.thin = 1,
  dea.model = "MCMCglmm",
  dea.nwarmup = 4096,
  dea.thin = 256,
  fdr.model = "ash",
  random.seed = 0,
  nthread = parallel::detectCores()\%/\%2
)
}
\arguments{
\item{component.deviations}{Set this to \code{TRUE} to do differential expression analysis on the component deviations as well as the group quants.}

\item{norm.model}{Either \code{NULL} (no normalisation), \code{"median"}, \code{"quantile"} or \code{"theta"} (seaMass-theta Bayesian normalisation)}

\item{norm.nwarmup}{Number of MCMC warmup iterations to run for each chain with seaMass-theta Bayesian normalisation.}

\item{norm.thin}{MCMC thinning factor with seaMass-theta Bayesian normalisation.}

\item{dea.model}{Either \code{NULL} (no differential expression analysis) or \code{"MCMCglmm"} (MCMCglmm differential expression analysis)}

\item{dea.nwarmup}{Number of MCMC warmup iterations to run for each chain with MCMCglmm differential expression analysis.}

\item{dea.thin}{MCMC thinning factor with MCMCglmm differential expression analysis.}

\item{fdr.model}{Either \code{NULL} (no false discovery rate correction) or \code{"ash"} (ash false discovery rate correction)}

\item{random.seed}{Random number seed}

\item{nthread}{Number of CPU threads to employ}
}
\description{
Define advanced control parameters for the seaMass-Σ Bayesian model.
}
\section{Functions}{
\itemize{
\item \code{delta_control}: Generator function
}}

