% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{execute_model}
\alias{execute_model}
\title{execute_model (internal)}
\usage{
execute_model(fit, block, chain, use.priors)
}
\arguments{
\item{fit}{deamass fit object.}

\item{chain}{Number of chain to process.}

\item{use.priors}{true or false}
}
\description{
execute_model (internal)
}
