% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{execute_model}
\alias{execute_model}
\title{execute_model (internal)}
\usage{
execute_model(fit, chain, priors = NULL)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{chain}{Number of chain to process.}

\item{priors}{The priors}
}
\description{
execute_model (internal)
}
