% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr_ash}
\alias{fdr_ash}
\title{False Discovery Rate correction with ash}
\usage{
fdr_ash(
  fit,
  input = "de",
  output = "fdr",
  by.effect = TRUE,
  by.model = TRUE,
  as.data.table = FALSE,
  use.df = TRUE,
  min.components = 1,
  min.components.per.condition = 0,
  min.measurements = 1,
  min.measurements.per.condition = 0,
  min.test.samples = 4,
  min.test.samples.per.condition = 2,
  min.real.samples = 1,
  min.real.samples.per.condition = 0,
  mixcompdist = "halfuniform",
  ...
)
}
\description{
False Discovery Rate correction with ash
}
