% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{fdr_BH}
\alias{fdr_BH}
\title{False Discovery Rate correction with the Benjamini-Hochberg method}
\usage{
fdr_BH(data, as.data.table = TRUE, ...)
}
\description{
False Discovery Rate correction with the Benjamini-Hochberg method
}
