% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.Progenesis.R
\name{import.Progenesis}
\alias{import.Progenesis}
\title{Add together two numbers.}
\usage{
import.Progenesis(datafile, only.used = T)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
