% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ProteomeDiscoverer.R
\name{import.ProteomeDiscoverer}
\alias{import.ProteomeDiscoverer}
\title{Import SCIEX ProteinPilot data.}
\usage{
import.ProteomeDiscoverer(file = NULL, fractions = NULL, data = NULL)
}
\arguments{
\item{file}{Location of the PeptideSummary.txt file.}

\item{fractions}{If your study involves more than one multiplex, `fractions` defines a mapping from the fraction index to a run ID.}

\item{data}{Advanced: Rather than specifying a `file`, you can enter an already loaded `data.frame` here.}
}
\value{
A `data.table` formatted for input into `bayesprot`.
}
\description{
`import.ProteinPilot` reads in a SCIEX ProteinPilot PeptideSummary.txt file for processing with BayesProt
}
