% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.ProteomeDiscoverer.R
\name{import.ProteomeDiscoverer}
\alias{import.ProteomeDiscoverer}
\title{Add together two numbers.}
\usage{
import.ProteomeDiscoverer(dd.raw, dd.fractions)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
