% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_MSstats}
\alias{import_MSstats}
\title{Import data outputed by an MSstats import routine}
\usage{
import_MSstats(data)
}
\arguments{
\item{data}{MSstats output
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{seamassdelta}.
}
\description{
Reads in a set of \code{_with_dscore} datasets processed by OpenSWATH and PyProphet for processing with \link{seamassdelta}.
}
