% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_OpenSwath}
\alias{import_OpenSwath}
\title{Import OpenSWATH data}
\usage{
import_OpenSwath(file = NULL, shared = FALSE, m_score.cutoff = 0.05,
  data = NULL)
}
\arguments{
\item{m_score.cutoff}{Include only measurements with PyProphet m_score >= than this?}

\item{data}{Advanced: Rather than specifying a \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}

\item{files}{A \code{csv} file to import.}
}
\value{
A \link{data.frame} for input into \link{seamassdelta}.
}
\description{
Reads in the output of an OpenSWATH -> PyProphet -> TRIC pipeline for processing with \link{seamassdelta}.
}
