% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ProteomeDiscoverer}
\alias{import_ProteomeDiscoverer}
\title{Import Thermo ProteomeDiscoverer data}
\usage{
import_ProteomeDiscoverer(file = NULL, shared = F, used = T,
  data = NULL)
}
\arguments{
\item{file}{Location of the \code{PSMs.txt} file.}

\item{shared}{Include shared components?}

\item{used}{Include only measurements marked as used by ProteomeDiscoverer?}

\item{data}{Advanced: Rather than specifying \code{file}, you can enter a \link{data.frame} preloaded with
\link[data.table]{fread} default parameters.}
}
\value{
A \link{data.frame} for input into \link{seamassdelta}.
}
\description{
Reads in a Thermo ProteomeDiscoverer \code{PSMs.txt} file for processing with \link{seamassdelta}.
}
