% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seamassdelta.R
\name{new_control}
\alias{new_control}
\title{Control parameters for the seamassdelta Bayesian model}
\usage{
new_control(measurement.model = "independent", measurement.eb.min = 2,
  component.model = NULL, component.eb.min = 2,
  assay.model = "independent", assay.eb.min = 2,
  error.model = "poisson", missingness.model = "censored",
  missingness.threshold = 0, model.seed = 0, model.nchain = 4,
  model.nwarmup = 256, model.thin = 1, model.nsample = 1024,
  squeeze.var.func = squeeze_var, dist.var.func = dist_invchisq_mcmc,
  dist.mean.func = dist_lst_mcmc,
  nthread = parallel::detectCores()\%/\%2, hpc = NULL)
}
\arguments{
\item{measurement.model}{Either \code{single} (single residual) or \code{independent} (per-measurement independent residuals; default)}

\item{measurement.eb.min}{Minimum number of measurements per component to use for computing Empirical Bayes priors}

\item{component.model}{Either \code{NULL} (no component model; default), \code{single} (single random effect) or \code{independent}
(per-component independent random effects)}

\item{component.eb.min}{Minimum number of components per group to use for computing Empirical Bayes priors}

\item{assay.model}{Either \code{NULL} (no assay model), \code{single} (single random effect) or \code{independent}
(per-assay independent random effects; default)}

\item{assay.eb.min}{Minimum number of assays per group group to use for computing Empirical Bayes priors}

\item{error.model}{Either \code{lognormal} or \code{poisson} (default)}

\item{missingness.model}{Either \code{zero} (NAs set to 0), \code{measurement} (NAs set to lowest quant of that measurement) or
\code{censored} (NAs modelled as censored between 0 and lowest quant of that measurement; default)}

\item{missingness.threshold}{All measurement quants equal to or below this are treated as missing (default = 0)}

\item{model.seed}{Random number seed}

\item{model.nchain}{Number of MCMC chains to run}

\item{model.nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{MCMC thinning factor}

\item{model.nsample}{Total number of MCMC samples to deliver downstream}

\item{nthread}{Number of CPU threads to employ}

\item{hpc}{Either \code{NULL} (execute locally), \code{pbs}, \code{sge} or \code{slurm} (submit to HPC cluster) [TODO]}
}
\value{
\code{seamassdelta_control} object to pass to \link{seamassdelta}
}
\description{
Control parameters for the seamassdelta Bayesian model
}
