% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm_median}
\alias{norm_median}
\title{Median normalisation}
\usage{
norm_median(fit, input = "standardised", output = "normalised",
  norm.groups = ".*", ...)
}
\description{
A Bayesian version of median normalisation.
}
