% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_fdr}
\alias{plot_fdr}
\title{Add together two numbers.}
\usage{
plot_fdr(data.fdr, y.max = NULL, y.col = "qvalue")
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
