% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca,seaMass-method}
\alias{plot_pca,seaMass-method}
\title{Robust PCA plot}
\usage{
\S4method{plot_pca}{seaMass}(
  object,
  data.summary = normalised_group_quants(object, summary = TRUE),
  data.mcmc = normalised_group_quants(object),
  data.design = assay_design(object),
  contours = 1:2,
  aspect.ratio = 9/16,
  robust = TRUE,
  ...
)
}
\arguments{
\item{object}{.}

\item{data.summary}{.}

\item{data.design}{.}

\item{contours}{.}

\item{robust}{.}

\item{data}{.}
}
\value{
A ggplot2 object.
}
\description{
Robust PCA plot
}
