% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_plots.R
\name{plot_pca,seaMass-method}
\alias{plot_pca,seaMass-method}
\title{Robust PCA plot}
\usage{
\S4method{plot_pca}{seaMass}(
  object,
  data.design = assay_design(object),
  variables = NULL,
  input = "model1",
  type = "standardised.group.deviations",
  scale = FALSE,
  robust = TRUE,
  contours = 1,
  aspect.ratio = 3/4,
  labels = 25,
  colour = "Condition",
  fill = "Condition",
  shape = "Condition",
  data = NULL,
  ...
)
}
\arguments{
\item{object}{.}

\item{data.design}{.}

\item{robust}{.}

\item{contours}{.}

\item{data}{.}

\item{data.summary}{.}
}
\value{
A ggplot2 object .
}
\description{
Robust PCA plot
}
