% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca,seaMass-method}
\alias{plot_pca,seaMass-method}
\title{Robust PCA plot}
\usage{
\S4method{plot_pca}{seaMass}(
  object,
  data.design = assay_design(object),
  type = "normalised.group.quants",
  robust = TRUE,
  contours = 1:2,
  aspect.ratio = 9/16,
  labels = TRUE,
  colour = "Condition",
  shape = NULL,
  ...
)
}
\arguments{
\item{object}{.}

\item{data.design}{.}

\item{robust}{.}

\item{contours}{.}

\item{data}{.}

\item{data.summary}{.}
}
\value{
A ggplot2 object.
}
\description{
Robust PCA plot
}
