% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Robust PCA plot}
\usage{
plot_pca(
  fit,
  data = normalised_group_quants(fit),
  data.summary = normalised_group_quants(fit, summary = T),
  data.design = assay_design(fit),
  contours = 1:2,
  robust = TRUE
)
}
\arguments{
\item{fit}{.}

\item{data}{.}

\item{data.summary}{.}

\item{data.design}{.}

\item{contours}{.}

\item{robust}{.}
}
\value{
A ggplot2 object.
}
\description{
Robust PCA plot
}
