% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Add together two numbers.}
\usage{
plot_pca(
  fit,
  data = normalised_group_quants(fit),
  data.summary = normalised_group_quants(fit, summary = T),
  data.design = assay_design(fit)
)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
