% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pca}
\alias{plot_pca}
\title{Add together two numbers.}
\usage{
plot_pca(fit, data = group_quants(fit, summary = F),
  data.summary = group_quants(fit), data.design = design(fit))
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
