% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_pr}
\alias{plot_pr}
\title{Precision-Recall plot}
\usage{
plot_pr(
  data.fdr,
  plot.fdr = TRUE,
  y.max = NULL,
  legend.nrow = 1,
  y.col = "qvalue"
)
}
\arguments{
\item{data.fdr}{.}

\item{y.max}{.}
}
\value{
A ggplot2 object.
}
\description{
Precision-Recall plot
}
