% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_pr}
\alias{plot_pr}
\title{Precision-Recall plot}
\usage{
plot_pr(data.fdr, plot.fdr = T, ymax = NULL, legend.nrow = 1)
}
\arguments{
\item{data.fdr}{.}

\item{ymax}{.}
}
\value{
A ggplot2 object.
}
\description{
Precision-Recall plot
}
