% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{plot_priors}
\alias{plot_priors}
\title{Add together two numbers.}
\usage{
plot_priors(
  data.fits = NULL,
  ci = c(0.05, 0.95),
  by = NULL,
  xlab = "v",
  ylim = NULL,
  trans = identity,
  inv.trans = identity,
  show.input = T
)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
