% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plots.R
\name{plot_volcano,seaMass_delta-method}
\alias{plot_volcano,seaMass_delta-method}
\title{Volcano plot}
\usage{
\S4method{plot_volcano}{seaMass_delta}(
  object,
  effect = NULL,
  error.bars = TRUE,
  stdev.col = "PosteriorSD",
  x.col = "PosteriorMean",
  y.col = "qvalue",
  width = 1024,
  height = 768,
  data.fdr = group_quants_fdr(object, effect),
  truth.func = NULL,
  output = "plotly",
  ggplot.nlabel = 25
)
}
\arguments{
\item{data.fdr}{.}
}
\value{
A ggplot2 object.
}
\description{
Volcano plot
}
