% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.de.R
\name{process.de}
\alias{process.de}
\title{Add together two numbers.}
\usage{
process.de()
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
