% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.input.R
\name{process.input}
\alias{process.input}
\title{Add together two numbers.}
\usage{
process.input(dd, id = "bayesprot", ref.assays = levels(dd$Assay),
  de.design = NULL, ...)
}
\arguments{
\item{datafile}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
