% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.input.R
\name{process.input}
\alias{process.input}
\title{process.input (BayesProt internal function)}
\usage{
process.input(dd, id, ref.assays, de.design, ...)
}
\arguments{
\item{dd}{The dataset returned by a BayesProt import() function.}

\item{...}{Other arguments in the bayesprot() or bayesprot.hpc() call}
}
\value{
BayesProt output directory structure created with the input populated in the 'input' sub-directory
}
\description{
process.input (BayesProt internal function)
}
