% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.model0.R
\name{process.model0}
\alias{process.model0}
\title{process.model0 (internal)}
\usage{
process.model0(chain)
}
\arguments{
\item{chain}{Number of chain to process.}
}
\description{
process.model0 (internal)
}
