% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.qprot.R
\name{process.qprot}
\alias{process.qprot}
\title{process.qprot (BayesProt internal function)}
\usage{
process.qprot(chain)
}
\arguments{
\item{chain}{.}
}
\value{
.
}
\description{
process.qprot (BayesProt internal function)
}
