% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.quant.R
\name{process.quant}
\alias{process.quant}
\title{process.quant (BayesProt internal function)}
\usage{
process.quant()
}
\arguments{
\item{input_dir}{.}
}
\value{
.
}
\description{
process.quant (BayesProt internal function)
}
