% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.study.R
\name{process.study}
\alias{process.study}
\title{process.study (BayesProt internal function)}
\usage{
process.study(input_dir)
}
\arguments{
\item{input_dir}{.}
}
\value{
.
}
\description{
process.study (BayesProt internal function)
}
