% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_model}
\alias{process_model}
\title{model (internal)}
\usage{
process_model(chain, path.results = ".", priors = NULL)
}
\arguments{
\item{chain}{Number of chain to process.}

\item{priors}{Priors for peptide and feature variances.}
}
\description{
model (internal)
}
