% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_model}
\alias{process_model}
\title{process_model (internal)}
\usage{
process_model(fit, block = 1, chain = 1)
}
\arguments{
\item{fit}{seamassdelta fit object.}

\item{chain}{Number of chain to process.}
}
\description{
process_model (internal)
}
