% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_model0}
\alias{process_model0}
\title{process_model0 (internal)}
\usage{
process_model0(fit, chain = 1)
}
\arguments{
\item{fit}{bayesprot fit object.}

\item{chain}{Number of chain to process.}
}
\description{
process_model0 (internal)
}
