% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_model0}
\alias{process_model0}
\title{process_model0 (internal)}
\usage{
process_model0(fit, block = 1, chain = 1)
}
\arguments{
\item{fit}{deamass fit object.}

\item{i}{.}
}
\description{
process_model0 (internal)
}
