% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model0.R
\name{process_model0}
\alias{process_model0}
\title{process_model0 (internal)}
\usage{
process_model0(chain, path.results = ".")
}
\arguments{
\item{chain}{Number of chain to process.}
}
\description{
process_model0 (internal)
}
