% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_model1}
\alias{process_model1}
\title{process_model1 (internal)}
\usage{
process_model1(chain, path.results = ".")
}
\arguments{
\item{chain}{Number of chain to process.}

\item{path.results}{Path to store results.}
}
\description{
process_model1 (internal)
}
