% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{process_plots}
\alias{process_plots}
\title{process_plots (internal)}
\usage{
process_plots(chain, fit = NULL)
}
\arguments{
\item{chain}{Number of chain to process.}

\item{fit}{bayesprot fit object.}
}
\description{
process_plots (internal)
}
