% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dea.R
\name{protein_de}
\alias{protein_de}
\title{Return differential expression as a list of FDR-controlled data tables}
\usage{
protein_de(fits, key = 1, as.data.table = F)
}
\description{
Return differential expression as a list of FDR-controlled data tables
}
