% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{runs}
\alias{runs}
\alias{runs<-}
\title{Imported dataset run information}
\usage{
runs(data)

runs(data) <- value
}
\arguments{
\item{data}{\link{data.frame} returned by \link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}.}

\item{value}{Edited \link{data.frame} created by \code{runs}.}
}
\description{
Get or set run information from a \link{data.frame} returned by \link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}.
Used to manually add run information to imported datasets
}
