% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule_sge-class}
\alias{schedule_sge-class}
\alias{schedule_sge}
\title{Schedule seaMass to run on SGE}
\usage{
schedule_sge(q = NULL, ppn = NULL, mem = NULL, walltime = NULL, M = NULL)
}
\arguments{
\item{ppn}{.}

\item{mem}{.}

\item{walltime}{.}

\item{M}{.}

\item{destination}{.}
}
\description{
Schedule seaMass to run on SGE
}
\section{Functions}{
\itemize{
\item \code{schedule_sge}: Generator function
}}

