% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule_sge-class}
\alias{schedule_sge-class}
\alias{schedule_sge}
\title{Schedule seaMass to run on SGE}
\usage{
schedule_sge(q = NULL, pe = NULL, mem = NULL, time = NULL, M = NULL)
}
\arguments{
\item{q}{.}

\item{pe}{.}

\item{mem}{.}

\item{time}{.}

\item{M}{.}
}
\description{
Schedule seaMass to run on SGE
}
\section{Functions}{
\itemize{
\item \code{schedule_sge}: Generator function
}}

