% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule_slurm-class}
\alias{schedule_slurm-class}
\alias{schedule_slurm}
\title{Schedule seaMass to run on SLURM}
\usage{
schedule_slurm(
  partition = NULL,
  cpus_per_task = NULL,
  mem = NULL,
  time = NULL,
  email = NULL
)
}
\arguments{
\item{partition}{.}

\item{mem}{.}

\item{time}{.}

\item{email}{.}

\item{nthread}{.}
}
\description{
Schedule seaMass to run on SLURM
}
\section{Functions}{
\itemize{
\item \code{schedule_slurm}: Generator function
}}

