% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass.R
\docType{class}
\name{seaMass-class}
\alias{seaMass-class}
\alias{normalised_group_quants,seaMass-method}
\title{seaMass object}
\usage{
\S4method{normalised_group_quants}{seaMass}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "normalised.group.quants",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)
}
\description{
Methods shared between \link{sigma_fit} and \link{seaMass_delta}
}
\section{Functions}{
\itemize{
\item \code{normalised_group_quants,seaMass-method}: Get the model normalised group quantifications as a \link{data.frame}.
}}

