% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_delta.R, R/sigma_block.R
\docType{class}
\name{seaMass_delta-class}
\alias{seaMass_delta-class}
\alias{seaMass_delta}
\alias{open_delta}
\alias{run,seaMass_delta-method}
\alias{name,seaMass_delta-method}
\alias{filepath,seaMass_delta-method}
\alias{parent,seaMass_delta-method}
\alias{control,seaMass_delta-method}
\alias{assay_design,seaMass_delta-method}
\alias{del,seaMass_delta-method}
\alias{group_quants_de,seaMass_delta-method}
\alias{component_deviations_de,seaMass_delta-method}
\alias{group_quants_fdr,seaMass_delta-method}
\alias{component_deviations_fdr,seaMass_delta-method}
\alias{add_seaMass_spikein_truth}
\alias{add_Navarro_spikein_truth}
\alias{group_means,sigma_block-method}
\title{seaMass-delta}
\usage{
seaMass_delta(
  fit,
  data.design = assay_design(fit),
  name = "default",
  control = delta_control(),
  ...
)

open_delta(fit, name = "default", quiet = FALSE, force = FALSE)

\S4method{run}{seaMass_delta}(object)

\S4method{name}{seaMass_delta}(object)

\S4method{filepath}{seaMass_delta}(object)

\S4method{parent}{seaMass_delta}(object)

\S4method{control}{seaMass_delta}(object)

\S4method{assay_design}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{del}{seaMass_delta}(object)

\S4method{group_quants_de}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "group.quants.de",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_deviations_de}{seaMass_delta}(
  object,
  components = NULL,
  summary = TRUE,
  type = "component.deviations.de",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_quants_fdr}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "group.quants.fdr",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_deviations_fdr}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "component.deviations.fdr",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

add_seaMass_spikein_truth(data.fdr)

add_Navarro_spikein_truth(data.fdr)

\S4method{group_means}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{A \link{seaMass_sigma} object as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-delta differential expression analysis and false discovery rate correction on normalised group-level
quants output by seaMass-theta
}
\section{Functions}{
\itemize{
\item \code{seaMass_delta()}: Runs seaMass-Σ.

\item \code{open_delta()}: Open a complete \code{seaMass_delta} run from the supplied \link{seaMass_sigma} fir object and \code{name}.

\item \code{run(seaMass_delta)}: Run.

\item \code{name(seaMass_delta)}: Get name.

\item \code{filepath(seaMass_delta)}: Get path.

\item \code{parent(seaMass_delta)}: Get the \link{seaMass_sigma} object.

\item \code{control(seaMass_delta)}: Get the \link{fit_control}.

\item \code{assay_design(seaMass_delta)}: Get the study design as a \code{data.frame}.

\item \code{del(seaMass_delta)}: Delete the \code{seaMass_delta} run from disk.

\item \code{group_quants_de(seaMass_delta)}: Get the model differential expression as a \link{data.frame}.

\item \code{component_deviations_de(seaMass_delta)}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{group_quants_fdr(seaMass_delta)}: Get the model differential expression as a \link{data.frame}.

\item \code{component_deviations_fdr(seaMass_delta)}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{add_seaMass_spikein_truth()}: Add our spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{add_Navarro_spikein_truth()}: Add Navarro spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{group_means(sigma_block)}: Get the model group means as a \link{data.frame}.

}}
