% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_delta.R
\docType{class}
\name{seaMass_delta-class}
\alias{seaMass_delta-class}
\alias{seaMass_delta}
\alias{open_seaMass_delta}
\alias{run,seaMass_delta-method}
\alias{name,seaMass_delta-method}
\alias{path,seaMass_delta-method}
\alias{control,seaMass_delta-method}
\alias{assay_design,seaMass_delta-method}
\alias{del,seaMass_delta-method}
\alias{components,seaMass_delta-method}
\alias{groups,seaMass_delta-method}
\alias{unnormalised_group_quants,seaMass_delta-method}
\alias{component_deviations,seaMass_delta-method}
\alias{group_variances,seaMass_delta-method}
\alias{group_de,seaMass_delta-method}
\alias{component_deviations_de,seaMass_delta-method}
\alias{group_fdr,seaMass_delta-method}
\alias{component_deviations_fdr,seaMass_delta-method}
\title{seaMass-Δ}
\usage{
seaMass_delta(
  sigma,
  data.design = assay_design(sigma),
  name = "fit",
  control = delta_control(),
  ...
)

open_seaMass_delta(sigma, name = "fit", quiet = FALSE, force = FALSE)

\S4method{run}{seaMass_delta}(object)

\S4method{name}{seaMass_delta}(object)

\S4method{path}{seaMass_delta}(object)

\S4method{control}{seaMass_delta}(object)

\S4method{assay_design}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{del}{seaMass_delta}(object)

\S4method{components}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{groups}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{unnormalised_group_quants}{seaMass_delta}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "standardised",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_deviations}{seaMass_delta}(
  object,
  components = NULL,
  summary = FALSE,
  input = "standardised.component.deviations",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{group_variances}{seaMass_delta}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "normalised.group.variances",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{group_de}{seaMass_delta}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "group.de",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_deviations_de}{seaMass_delta}(
  object,
  components = NULL,
  summary = FALSE,
  input = "component.deviations.de",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{group_fdr}{seaMass_delta}(object, input = "group.fdr", as.data.table = FALSE)

\S4method{component_deviations_fdr}{seaMass_delta}(
  object,
  input = "component.deviations.fdr",
  as.data.table = FALSE
)
}
\arguments{
\item{sigma}{A \link{seaMass_sigma} object as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on unnormalised group-level
quants output by seaMass-Σ
}
\section{Functions}{
\itemize{
\item \code{seaMass_delta}: Runs seaMass-Σ.

\item \code{open_seaMass_delta}: Open a complete \code{seaMass_delta} run from the supplied \link{seaMass_sigma} and \code{name}.

\item \code{run,seaMass_delta-method}: Run.

\item \code{name,seaMass_delta-method}: Get name.

\item \code{path,seaMass_delta-method}: Get path.

\item \code{control,seaMass_delta-method}: Get the \link{sigma_control}.

\item \code{assay_design,seaMass_delta-method}: Get the study design as a \code{data.frame}.

\item \code{del,seaMass_delta-method}: Delete the \code{seaMass_delta} run from disk.

\item \code{components,seaMass_delta-method}: Get the component metadata as a \link{data.frame}.

\item \code{groups,seaMass_delta-method}: Get the group metadata as a \link{data.frame}.

\item \code{unnormalised_group_quants,seaMass_delta-method}: Get the model unnormalised group quantifications as a \link{data.frame}.

\item \code{component_deviations,seaMass_delta-method}: Get the input standardised component deviations as a \link{data.frame}.

\item \code{group_variances,seaMass_delta-method}: Get the model normalised group variances as a \link{data.frame}.

\item \code{group_de,seaMass_delta-method}: Get the model differential expression as a \link{data.frame}.

\item \code{component_deviations_de,seaMass_delta-method}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{group_fdr,seaMass_delta-method}: Get the false discovery rate corrected group differential expression as a \link{data.frame}.

\item \code{component_deviations_fdr,seaMass_delta-method}: Get the false discovery rate corrected component deviations differential expression as a \link{data.frame}.
}}

