% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_delta.R, R/sigma_block.R
\docType{class}
\name{seaMass_delta-class}
\alias{seaMass_delta-class}
\alias{seaMass_delta}
\alias{run,seaMass_delta-method}
\alias{name,seaMass_delta-method}
\alias{root,seaMass_delta-method}
\alias{filepath,seaMass_delta-method}
\alias{parent,seaMass_delta-method}
\alias{control,seaMass_delta-method}
\alias{assay_design,seaMass_delta-method}
\alias{del,seaMass_delta-method}
\alias{group_quants_de,seaMass_delta-method}
\alias{component_deviations_de,seaMass_delta-method}
\alias{group_quants_fdr,seaMass_delta-method}
\alias{component_deviations_fdr,seaMass_delta-method}
\alias{truth_seaMass_spikein}
\alias{truth_Navarro_spikein}
\alias{group_means,sigma_block-method}
\title{seaMass-delta}
\usage{
seaMass_delta(
  fit,
  data.design = seaMass::assay_design(fit),
  name = seaMass::name(fit),
  control = seaMass::delta_control(),
  ...
)

\S4method{run}{seaMass_delta}(object)

\S4method{name}{seaMass_delta}(object)

\S4method{root}{seaMass_delta}(object)

\S4method{filepath}{seaMass_delta}(object)

\S4method{parent}{seaMass_delta}(object)

\S4method{control}{seaMass_delta}(object)

\S4method{assay_design}{seaMass_delta}(object, as.data.table = FALSE)

\S4method{del}{seaMass_delta}(object)

\S4method{group_quants_de}{seaMass_delta}(
  object,
  groups = NULL,
  summary = TRUE,
  type = "group.quants.de",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{component_deviations_de}{seaMass_delta}(
  object,
  components = NULL,
  summary = TRUE,
  type = "component.deviations.de",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_quants_fdr}{seaMass_delta}(
  object,
  groups = NULL,
  type = "group.quants.fdr",
  as.data.table = FALSE
)

\S4method{component_deviations_fdr}{seaMass_delta}(
  object,
  groups = NULL,
  type = "component.deviations.fdr",
  as.data.table = FALSE
)

truth_seaMass_spikein(data.fdr)

truth_Navarro_spikein(data.fdr)

\S4method{group_means}{sigma_block}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{A \link{seaMass_sigma} object as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-delta differential expression analysis and false discovery rate correction on normalised group-level
quants output by seaMass-theta
}
\section{Functions}{
\itemize{
\item \code{seaMass_delta}: Runs seaMass-Σ.

\item \code{run,seaMass_delta-method}: Run.

\item \code{name,seaMass_delta-method}: Get name.

\item \code{root,seaMass_delta-method}: Get the \code{seaMass_sigma} object.

\item \code{filepath,seaMass_delta-method}: Get path.

\item \code{parent,seaMass_delta-method}: Get the \link{seaMass_sigma} object.

\item \code{control,seaMass_delta-method}: Get the \link{fit_control}.

\item \code{assay_design,seaMass_delta-method}: Get the study design as a \code{data.frame}.

\item \code{del,seaMass_delta-method}: Delete the \code{seaMass_delta} run from disk.

\item \code{group_quants_de,seaMass_delta-method}: Get the model differential expression as a \link{data.frame}.

\item \code{component_deviations_de,seaMass_delta-method}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{group_quants_fdr,seaMass_delta-method}: Get the model differential expression as a \link{data.frame}.

\item \code{component_deviations_fdr,seaMass_delta-method}: Get the component deviations differential expression as a \link{data.frame}.

\item \code{truth_seaMass_spikein}: Add our spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{truth_Navarro_spikein}: Add Navarro spikein ground truth to the object of \code{group_fdr} pr \code{component_deviations_fdr}.

\item \code{group_means,sigma_block-method}: Get the model group means as a \link{data.frame}.
}}

