% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_fit.R
\name{seaMass_delta}
\alias{seaMass_delta}
\alias{open_delta_fit}
\alias{del.seaMass_delta_fit}
\alias{control.seaMass_delta_fit}
\alias{design.seaMass_delta_fit}
\alias{groups.seaMass_delta_fit}
\alias{unnormalised_group_quants.seaMass_delta_fit}
\alias{normalised_group_quants.seaMass_delta_fit}
\title{seaMass-Δ}
\usage{
seaMass_delta(
  fits,
  data.design = design(fits),
  ref.groups = levels(groups(fit)$Group),
  summaries = FALSE,
  plots = FALSE,
  name = sub("^(.*)\\\\..*\\\\.seaMass-sigma", "\\\\1", basename(fits[[1]])),
  control = new_delta_control()
)

open_delta_fit(dir, quiet = FALSE, force = FALSE)

\method{del}{seaMass_delta_fit}(fit)

\method{control}{seaMass_delta_fit}(fit)

\method{design}{seaMass_delta_fit}(fit, as.data.table = F)

\method{groups}{seaMass_delta_fit}(fit, as.data.table = FALSE)

\method{unnormalised_group_quants}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary.func = dist_lst_mcmc,
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)

\method{normalised_group_quants}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary.func = dist_lst_mcmc,
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fits}{A list of \link{seaMass_sigma_fit} objects, as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{summaries}{Generate all summaries}

\item{plots}{Generate all plots}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta_fit} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on unnormalised group-level
quants output by seaMass-Σ
}
\section{Methods (by generic)}{
\itemize{
\item \code{del}: Delete the directory backing a previously created \code{seaMass_delta} object.

\item \code{control}: Returns the \link{seaMass_delta_control} object from an open \code{seaMass_delta_fit} object.

\item \code{design}: Returns the study design \code{data.frame} from an open \code{seaMass_delta_fit} object.

\item \code{groups}: Returns the group metadata from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{unnormalised_group_quants}: Returns the input unnormalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{normalised_group_quants}: Returns the normalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.
}}

