% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_fit.R, R/sigma_fit.R
\name{seaMass_delta}
\alias{seaMass_delta}
\alias{open_delta_fit}
\alias{del.seaMass_delta_fit}
\alias{control.seaMass_delta_fit}
\alias{sigma_fits.seaMass_delta_fit}
\alias{assay_design.seaMass_delta_fit}
\alias{groups.seaMass_delta_fit}
\alias{unnormalised_group_quants.seaMass_delta_fit}
\alias{normalised_group_quants.seaMass_delta_fit}
\alias{group_vars.seaMass_delta_fit}
\alias{group_de.seaMass_delta_fit}
\alias{group_fdr.seaMass_delta_fit}
\alias{unnormalised_group_quants.seaMass_sigma_fit}
\title{seaMass-Δ}
\usage{
seaMass_delta(
  sigma_fits,
  data.design = assay_design(sigma_fits),
  norm.groups = ".*",
  name = sub("^(.*)\\\\..*\\\\.seaMass-sigma", "\\\\1", basename(sigma_fits[[1]])),
  control = new_delta_control(),
  ...
)

open_delta_fit(name, quiet = FALSE, force = FALSE)

\method{del}{seaMass_delta_fit}(fit)

\method{control}{seaMass_delta_fit}(fit)

\method{sigma_fits}{seaMass_delta_fit}(fit)

\method{assay_design}{seaMass_delta_fit}(fit, as.data.table = F)

\method{groups}{seaMass_delta_fit}(fit, as.data.table = FALSE)

\method{unnormalised_group_quants}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "standardised",
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)

\method{normalised_group_quants}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "normalised",
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)

\method{group_vars}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "normalised.group.vars",
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)

\method{group_de}{seaMass_delta_fit}(
  fit,
  groups = NULL,
  summary = FALSE,
  input = "de",
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)

\method{group_fdr}{seaMass_delta_fit}(fit, input = "fdr", as.data.table = FALSE)

\method{unnormalised_group_quants}{seaMass_sigma_fit}(
  fit,
  groups = NULL,
  summary = FALSE,
  dist_lst_mcmc,
  input = "model1",
  chains = 1:control(fit)$model.nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{sigma_fits}{A list of \link{seaMass_sigma_fit} objects, as returned by seaMass-Σ.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
sample names, RefWeight channels, and any covariates specified by the experimental design.}

\item{name}{Name of folder on disk where all intermediate and output data will be stored; default is \code{"output"}.}

\item{control}{A control object created with \link{new_delta_control} specifying control parameters for the differential analysis.}
}
\value{
A \code{seaMass_delta_fit} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-Δ normalisation, differential expression analysis and/or false discovery rate correction on unnormalised group-level
quants output by seaMass-Σ
}
\section{Methods (by generic)}{
\itemize{
\item \code{del}: Delete the directory backing a previously created \code{seaMass_delta} object.

\item \code{control}: Returns the \link{seaMass_delta_control} object from an open \code{seaMass_delta_fit} object.

\item \code{sigma_fits}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{assay_design}: Returns the study design \code{data.frame} from an open \code{seaMass_delta_fit} object.

\item \code{groups}: Returns the group metadata from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{unnormalised_group_quants}: Returns the input unnormalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{normalised_group_quants}: Returns the normalised group quantifications from an open \code{seaMass_delta_fit} object as a \code{data.frame}.

\item \code{group_vars}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{group_de}: Returns the differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{group_fdr}: Returns the false discovery rate correct differential expression from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{unnormalised_group_quants}: Returns the model unnormalised group quantifications from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.
}}

