% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_sigma.R, R/sigma_block.R
\docType{class}
\name{seaMass_sigma-class}
\alias{seaMass_sigma-class}
\alias{seaMass_sigma}
\alias{open_sigma}
\alias{del,seaMass_sigma-method}
\alias{name,seaMass_sigma-method}
\alias{filepath,seaMass_sigma-method}
\alias{run,seaMass_sigma-method}
\alias{control,seaMass_sigma-method}
\alias{groups,seaMass_sigma-method}
\alias{components,seaMass_sigma-method}
\alias{assay_groups,seaMass_sigma-method}
\alias{assay_components,seaMass_sigma-method}
\alias{measurements,seaMass_sigma-method}
\alias{blocks,seaMass_sigma-method}
\alias{open_deltas,seaMass_sigma-method}
\alias{assay_design,seaMass_sigma-method}
\alias{timings,seaMass_sigma-method}
\alias{assay_means,seaMass_sigma-method}
\alias{assay_stdevs,seaMass_sigma-method}
\alias{assay_deviations,seaMass_sigma-method}
\alias{measurement_means,seaMass_sigma-method}
\alias{measurement_stdevs,seaMass_sigma-method}
\alias{component_means,seaMass_sigma-method}
\alias{component_stdevs,seaMass_sigma-method}
\alias{component_deviations,seaMass_sigma-method}
\alias{group_quants,seaMass_sigma-method}
\alias{group_means,seaMass_sigma-method}
\alias{normalised_group_means,seaMass_sigma-method}
\alias{normalised_group_stdevs,seaMass_sigma-method}
\alias{normalised_group_quants,seaMass_sigma-method}
\alias{standardised_group_deviations,seaMass_sigma-method}
\alias{name,sigma_block-method}
\title{seaMass-Σ}
\usage{
seaMass_sigma(
  data,
  data.design = new_assay_design(data),
  path = "fit",
  run = TRUE,
  control = sigma_control(),
  ...
)

open_sigma(path = "fit.seaMass", quiet = FALSE, force = FALSE)

\S4method{del}{seaMass_sigma}(object)

\S4method{name}{seaMass_sigma}(object)

\S4method{filepath}{seaMass_sigma}(object)

\S4method{run}{seaMass_sigma}(object)

\S4method{control}{seaMass_sigma}(object)

\S4method{groups}{seaMass_sigma}(object, summary = FALSE, as.data.table = FALSE)

\S4method{components}{seaMass_sigma}(object, summary = FALSE, as.data.table = FALSE)

\S4method{assay_groups}{seaMass_sigma}(object, summary = FALSE, as.data.table = FALSE)

\S4method{assay_components}{seaMass_sigma}(object, summary = FALSE, as.data.table = FALSE)

\S4method{measurements}{seaMass_sigma}(object, summary = FALSE, as.data.table = FALSE)

\S4method{blocks}{seaMass_sigma}(object)

\S4method{open_deltas}{seaMass_sigma}(object, quiet = FALSE, force = FALSE)

\S4method{assay_design}{seaMass_sigma}(object, as.data.table = FALSE)

\S4method{timings}{seaMass_sigma}(object, input = "model1", as.data.table = FALSE)

\S4method{assay_means}{seaMass_sigma}(
  object,
  assays = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{assay_stdevs}{seaMass_sigma}(object, input = "model1", as.data.table = FALSE)

\S4method{assay_deviations}{seaMass_sigma}(
  object,
  assays = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{measurement_means}{seaMass_sigma}(
  object,
  measurements = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{measurement_stdevs}{seaMass_sigma}(
  object,
  measurements = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_means}{seaMass_sigma}(
  object,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_stdevs}{seaMass_sigma}(
  object,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_deviations}{seaMass_sigma}(
  object,
  components = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{group_quants}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{group_means}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{normalised_group_means}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{normalised_group_stdevs}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{normalised_group_quants}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{standardised_group_deviations}{seaMass_sigma}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{name}{sigma_block}(object)
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_MaxQuant}, \link{import_OpenSWATH},
\link{import_ProteinPilot} or \link{import_ProteomeDiscoverer}, .}

\item{data.design}{A \link{data.frame} created by \link{new_assay_design} and then customised, which specifies
assay names and block design.}

\item{path}{Name of folder prefix on disk where all intermediate and output data will be stored.}

\item{run}{Run seaMass-Σ now, or just prepare it for later execution on e.g. a HPC cluster?}

\item{control}{A control object created with \link{sigma_control} specifying control parameters for the model.}
}
\value{
A \code{seaMass_sigma} object, which allows access to metadata and each block's \link{sigma_block} object to access
  various results.
}
\description{
Fits the seaMass-Σ Bayesian group-level quantification model to imported data.
}
\section{Functions}{
\itemize{
\item \code{seaMass_sigma}: Runs seaMass-Σ.

\item \code{open_sigma}: Open a complete \code{seaMass_sigma} run from the supplied \code{path}.

\item \code{del,seaMass_sigma-method}: Delete the \code{seaMass_sigma} run from disk.

\item \code{name,seaMass_sigma-method}: Get name.

\item \code{filepath,seaMass_sigma-method}: Get path.

\item \code{run,seaMass_sigma-method}: Run.

\item \code{control,seaMass_sigma-method}: Get the \link{sigma_control}.

\item \code{groups,seaMass_sigma-method}: Get the groups as a \code{data.frame}.

\item \code{components,seaMass_sigma-method}: Get the groups as a \code{data.frame}.

\item \code{assay_groups,seaMass_sigma-method}: Get the assay groups as a \code{data.frame}.

\item \code{assay_components,seaMass_sigma-method}: Get the assay groups as a \code{data.frame}.

\item \code{measurements,seaMass_sigma-method}: Get the groups as a \code{data.frame}.

\item \code{blocks,seaMass_sigma-method}: Get the list of \link{sigma_block} obejcts for the blocks.

\item \code{open_deltas,seaMass_sigma-method}: Open the list of \link{seaMass_delta} objects.

\item \code{assay_design,seaMass_sigma-method}: Get the study design as a \code{data.frame}.

\item \code{timings,seaMass_sigma-method}: Get the model timings as a \link{data.frame}.

\item \code{assay_means,seaMass_sigma-method}: Get the model assay means as a \link{data.frame}.

\item \code{assay_stdevs,seaMass_sigma-method}: Get the model assay stdevs as a \link{data.frame}.

\item \code{assay_deviations,seaMass_sigma-method}: Get the model assay deviations as a \link{data.frame}.

\item \code{measurement_means,seaMass_sigma-method}: Get the model measurement means as a \link{data.frame}.

\item \code{measurement_stdevs,seaMass_sigma-method}: Get the model measurement stdevs as a \link{data.frame}.

\item \code{component_means,seaMass_sigma-method}: Get the model component means as a \link{data.frame}.

\item \code{component_stdevs,seaMass_sigma-method}: Get the model component stdevs as a \link{data.frame}.

\item \code{component_deviations,seaMass_sigma-method}: Gets the model component deviations as a \link{data.frame}.

\item \code{group_quants,seaMass_sigma-method}: Get the model group quantifications as a \link{data.frame}.

\item \code{group_means,seaMass_sigma-method}: Get the model group means as a \link{data.frame}.

\item \code{normalised_group_means,seaMass_sigma-method}: Get the model normalised group stdevs as a \link{data.frame}.

\item \code{normalised_group_stdevs,seaMass_sigma-method}: Get the model normalised group stdevs as a \link{data.frame}.

\item \code{normalised_group_quants,seaMass_sigma-method}: Get the model normalised group quantifications as a \link{data.frame}.

\item \code{standardised_group_deviations,seaMass_sigma-method}: Get the model standardised group deviations as a \link{data.frame}.

\item \code{name,sigma_block-method}: Get the block name.
}}

