% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R, R/sigma_fit.R
\name{seaMass_sigma}
\alias{seaMass_sigma}
\alias{open_sigma_fits}
\alias{del.seaMass_sigma_fit}
\alias{control.seaMass_sigma_fit}
\alias{assay_design.seaMass_sigma_fit}
\alias{assay_design.seaMass_sigma_fits}
\alias{measurements.seaMass_sigma_fit}
\alias{components.seaMass_sigma_fit}
\alias{groups.seaMass_sigma_fit}
\alias{summary.seaMass_sigma_fit}
\alias{timings.seaMass_sigma_fit}
\alias{measurement_vars.seaMass_sigma_fit}
\alias{component_vars.seaMass_sigma_fit}
\alias{assay_vars.seaMass_sigma_fit}
\alias{component_deviations.seaMass_sigma_fit}
\alias{assay_deviations.seaMass_sigma_fit}
\title{seaMass-Σ}
\usage{
seaMass_sigma(data, data.design = new_assay_design(data),
  summaries = FALSE, plots = FALSE, name = "fit",
  control = new_sigma_control(), hpc.schedule = new_hpc_schedule())

open_sigma_fits(names, quiet = FALSE, force = FALSE)

\method{del}{seaMass_sigma_fit}(fit)

\method{control}{seaMass_sigma_fit}(fit)

\method{assay_design}{seaMass_sigma_fit}(fit, as.data.table = F)

\method{assay_design}{seaMass_sigma_fits}(fits, as.data.table = F)

\method{measurements}{seaMass_sigma_fit}(fit, as.data.table = FALSE)

\method{components}{seaMass_sigma_fit}(fit, as.data.table = FALSE)

\method{groups}{seaMass_sigma_fit}(fit, as.data.table = FALSE)

\method{summary}{seaMass_sigma_fit}(fit, group, input = "model1")

\method{timings}{seaMass_sigma_fit}(fit, input = "model1",
  as.data.table = FALSE)

\method{measurement_vars}{seaMass_sigma_fit}(fit, measurements = NULL,
  summary = FALSE, input = "model1",
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

\method{component_vars}{seaMass_sigma_fit}(fit, components = NULL,
  summary = FALSE, input = "model1",
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

\method{assay_vars}{seaMass_sigma_fit}(fit, groups = NULL,
  summary = FALSE, input = "model1",
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

\method{component_deviations}{seaMass_sigma_fit}(fit, components = NULL,
  summary = FALSE, input = "model1",
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)

\method{assay_deviations}{seaMass_sigma_fit}(fit, measurements = NULL,
  components = NULL, summary = FALSE, input = "model1",
  chains = 1:control(fit)$model.nchain, as.data.table = FALSE)
}
\arguments{
\item{data}{A \link{data.frame} of input data as returned by \link{import_ProteinPilot}, \link{import_ProteinPilot},
\link{import_ProteomeDiscovery}, \link{import_Progenesis} or \link{import_OpenSWATH}.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies
assay names and block design.}

\item{summaries}{Generate all summaries.}

\item{plots}{Generate all plots.}

\item{name}{Name of folder prefix on disk where all intermediate and output data will be stored; default is \code{"fit"}.}

\item{control}{A control object created with \link{new_sigma_control} specifying control parameters for the model.}

\item{hpc.schedule}{A hpc object created with \link{new_hpc_schedule} specifying hpc parameters for the type of HPC system seaMass will be deployed on.}
}
\value{
A \code{seaMass_sigma_fits} object, which is a list of \code{seaMass_sigma_fit} objects that can be interrogated
  for various metadata and results.
}
\description{
Fit the seaMass-Σ Bayesian group-level quantification model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{del}: Delete the directories backing a previously created list of \code{seaMass_sigma_fit} objects.

\item \code{control}: Returns the \link{seaMass_sigma_control} object from an open \code{seaMass_sigma_fit} object.

\item \code{assay_design}: Returns the study design \code{data.frame} from an open \code{seaMass_sigma_fit} object.

\item \code{assay_design}: Returns the full study design \code{data.frame} from an open \code{seaMass_sigma_fits} object.

\item \code{measurements}: Returns the measurement metadata from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{components}: Returns the component metadata from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{groups}: Returns the group metadata from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{summary}: Returns the model summary for a groupID from an open \code{seaMass_sigma_fit} object.

\item \code{timings}: Returns the model timings from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{measurement_vars}: Returns the model measurement variances from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{component_vars}: Returns the model component variances from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{assay_vars}: Returns the model assay variances from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{component_deviations}: Returns the model component deviations from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.

\item \code{assay_deviations}: Returns the model assay deviations from an open \code{seaMass_sigma_fit} object as a \code{data.frame}.
}}

