% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaMass_delta.R, R/seaMass_theta.R
\docType{class}
\name{root,seaMass_delta-method}
\alias{root,seaMass_delta-method}
\alias{seaMass_theta-class}
\alias{seaMass_theta}
\alias{run,seaMass_theta-method}
\alias{name,seaMass_theta-method}
\alias{root,seaMass_theta-method}
\alias{filepath,seaMass_theta-method}
\alias{parent,seaMass_theta-method}
\alias{del,seaMass_theta-method}
\alias{control,seaMass_theta-method}
\alias{assay_design,seaMass_theta-method}
\alias{assay_means,seaMass_theta-method}
\alias{group_standards,seaMass_theta-method}
\alias{group_quants,seaMass_theta-method}
\title{seaMass-theta}
\usage{
\S4method{root}{seaMass_delta}(object)

seaMass_theta(
  fit,
  data.design = assay_design(fit),
  norm.groups = top_groups(fit),
  name = "default",
  control = theta_control(),
  ...
)

\S4method{run}{seaMass_theta}(object)

\S4method{name}{seaMass_theta}(object)

\S4method{root}{seaMass_theta}(object)

\S4method{filepath}{seaMass_theta}(object)

\S4method{parent}{seaMass_theta}(object)

\S4method{del}{seaMass_theta}(object)

\S4method{control}{seaMass_theta}(object)

\S4method{assay_design}{seaMass_theta}(object, as.data.table = FALSE)

\S4method{assay_means}{seaMass_theta}(
  object,
  assays = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_standards}{seaMass_theta}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)

\S4method{group_quants}{seaMass_theta}(
  object,
  groups = NULL,
  summary = TRUE,
  input = "model1",
  chains = 1:control(object)@nchain,
  as.data.table = FALSE
)
}
\arguments{
\item{fit}{A \link{seaMass_sigma} object as returned by seaMass-sigma.}

\item{data.design}{Optionally, a \link{data.frame} created by \link{new_design} and then customised, which specifies RefWeight channels.}

\item{name}{Name of subfolder on disk where all intermediate and output data will be stored; default is \code{"default"}.}

\item{control}{A control object created with \link{new_theta_control} specifying control parameters for the normalisation.}

\item{groups}{Groups to normalise on, default is the top 10 that are in every block (instead of all, just for computational considerations).}
}
\value{
A \code{seaMass_theta} object that can be interrogated for various metadata and results.
}
\description{
Perform seaMass-theta normalisation on raw group-level quants output by seaMass-sigma
}
\section{Functions}{
\itemize{
\item \code{root,seaMass_delta-method}: Get the \code{seaMass_sigma} object.

\item \code{seaMass_theta}: Runs seaMass-theta.

\item \code{run,seaMass_theta-method}: Run.

\item \code{name,seaMass_theta-method}: Get name.

\item \code{root,seaMass_theta-method}: Get the \code{seaMass_sigma} object.

\item \code{filepath,seaMass_theta-method}: Get path.

\item \code{parent,seaMass_theta-method}: Get the \link{seaMass_sigma} object.

\item \code{del,seaMass_theta-method}: Delete the \code{seaMass_theta} run from disk.

\item \code{control,seaMass_theta-method}: Get the \link{fit_control}.

\item \code{assay_design,seaMass_theta-method}: Get the study design as a \code{data.frame}.

\item \code{assay_means,seaMass_theta-method}: Get the model assay means as a \link{data.frame}.

\item \code{group_standards,seaMass_theta-method}: Get the model group standards as a \link{data.frame}.

\item \code{group_quants,seaMass_theta-method}: Get the model group quantifications as a \link{data.frame}.
}}

