% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_block.R
\docType{class}
\name{sigma_block-class}
\alias{sigma_block-class}
\alias{sigma_block}
\alias{filepath,sigma_block-method}
\alias{control,sigma_block-method}
\alias{parent,sigma_block-method}
\alias{blocks,sigma_block-method}
\alias{groups,sigma_block-method}
\alias{measurements,sigma_block-method}
\alias{components,sigma_block-method}
\alias{assay_design,sigma_block-method}
\alias{priors,sigma_block-method}
\alias{summary,sigma_block-method}
\alias{timings,sigma_block-method}
\alias{assay_variances,sigma_block-method}
\alias{measurement_variances,sigma_block-method}
\alias{component_variances,sigma_block-method}
\alias{component_deviations,sigma_block-method}
\alias{assay_deviations,sigma_block-method}
\alias{raw_group_quants,sigma_block-method}
\alias{assay_exposures,sigma_block-method}
\alias{normalised_group_variances,sigma_block-method}
\title{seaMass-Σ block}
\usage{
\S4method{filepath}{sigma_block}(object)

\S4method{control}{sigma_block}(object)

\S4method{parent}{sigma_block}(object)

\S4method{blocks}{sigma_block}(object)

\S4method{groups}{sigma_block}(object, as.data.table = FALSE)

\S4method{measurements}{sigma_block}(object, as.data.table = FALSE)

\S4method{components}{sigma_block}(object, as.data.table = FALSE)

\S4method{assay_design}{sigma_block}(object, as.data.table = FALSE)

\S4method{priors}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{summary}{sigma_block}(object, group, chain, input = "model1")

\S4method{timings}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{assay_variances}{sigma_block}(object, input = "model1", as.data.table = FALSE)

\S4method{measurement_variances}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_variances}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{component_deviations}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{assay_deviations}{sigma_block}(
  object,
  assays = NULL,
  summary = FALSE,
  input = "model0",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{raw_group_quants}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{assay_exposures}{sigma_block}(
  object,
  assays = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)

\S4method{normalised_group_variances}{sigma_block}(
  object,
  groups = NULL,
  summary = FALSE,
  input = "model1",
  chains = 1:control(object)@model.nchain,
  as.data.table = FALSE
)
}
\description{
The results of a seaMass-Σ block to a single block, returned by /link{seaMass_sigma} /code{blocks()} function
}
\section{Methods (by generic)}{
\itemize{
\item \code{filepath}: Get the path.

\item \code{control}: Get the \link{sigma_control} object for this block.

\item \code{parent}: Get the \link{seaMass_sigma} object for this block.

\item \code{blocks}: Get the list of \link{sigma_block} obejcts for the blocks.

\item \code{groups}: Get the group metadata as a \link{data.frame}.

\item \code{measurements}: Get the measurement metadata as a \link{data.frame}.

\item \code{components}: Get the component metadata as a \link{data.frame}.

\item \code{assay_design}: Get the study design for this block as a \link{data.frame}.

\item \code{priors}: Get the priors if computed.

\item \code{summary}: Print the model summary for a group.

\item \code{timings}: Get the model timings as a \link{data.frame}.

\item \code{assay_variances}: Get the model assay variances as a \link{data.frame}.

\item \code{measurement_variances}: Get the model measurement variances as a \link{data.frame}.

\item \code{component_variances}: Get the model component variances as a \link{data.frame}.

\item \code{component_deviations}: Gets the model component deviations as a \link{data.frame}.

\item \code{assay_deviations}: Get the model assay deviations as a \link{data.frame}.

\item \code{raw_group_quants}: Get the model raw group quantifications as a \link{data.frame}.

\item \code{assay_exposures}: Get the model assay exposures as a \link{data.frame}.

\item \code{normalised_group_variances}: Get the model normalised group variances as a \link{data.frame}.
}}

