% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_control.R
\docType{class}
\name{sigma_control-class}
\alias{sigma_control-class}
\alias{sigma_control}
\title{Control parameters for seaMass-Σ}
\usage{
sigma_control(
  summarise = NULL,
  keep = c("measurement.variances", "component.variances", "component.deviations",
    "unnormalised.group.quants"),
  plot = NULL,
  measurement.model = "independent",
  measurement.eb.min = 2,
  component.model = "independent",
  component.eb.min = 3,
  assay.model = "component",
  assay.eb.min = 3,
  assay.eb.nsample = 16,
  error.model = "lognormal",
  missingness.model = "censored",
  missingness.threshold = 0,
  model.nchain = 4,
  model.nwarmup = 256,
  model.thin = 4,
  model.nsample = 1024,
  eb.max = 1024,
  random.seed = 0,
  nthread = parallel::detectCores()\%/\%2,
  schedule = schedule_local()
)
}
\arguments{
\item{summarise}{Outputs to write csv summaries for, \code{NULL} or a subset of
\code{c("measurement.variances", "component.variances", "component.deviations", "assay.deviations", "unnormalised.group.quants")}}

\item{keep}{Outputs to keep MCMC samples for, \code{NULL} or a subset of
\code{c("measurement.variances", "component.variances", "component.deviations", "assay.deviations", "unnormalised.group.quants")}
Note, you must keep \code{"unnormalised.group.quants"} if you want to run seaMass-Δ!}

\item{plot}{Outputs to plot [TODO]}

\item{measurement.model}{Either \code{"single"} (single residual) or \code{"independent"} (per-measurement independent residuals)}

\item{measurement.eb.min}{Minimum number of measurements per component to use for computing Empirical Bayes priors}

\item{component.model}{Either \code{NULL} (no component model), \code{"single"} (single random effect) or \code{"independent"}
(per-component independent random effects; default)}

\item{component.eb.min}{Minimum number of components per group to use for computing Empirical Bayes priors}

\item{assay.model}{Either \code{NULL} (no assay model), \code{"measurement"} (per-assay independent random effects across measurements)
or \code{"componenet"} (per-assay independent random effects across components; default)}

\item{assay.eb.min}{Minimum number of assays per group group to use for computing Empirical Bayes priors}

\item{assay.eb.nsample}{Number of MCMC samples to use for assay model input}

\item{error.model}{Likelihood model, either \code{"poisson"} or \code{"lognormal"} (default)}

\item{missingness.model}{Either \code{NULL} (NAs removed), \code{"one"} (NAs set to 1), \code{"minimum"} (NAs set to lowest quant of that measurement) or
\code{"censored"} (NAs modelled as censored below lowest quant of that measurement; default)}

\item{missingness.threshold}{All datapoints equal to or below this count are treated as missing}

\item{model.nchain}{Number of MCMC chains to run}

\item{model.nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{model.thin}{MCMC thinning factor}

\item{model.nsample}{Total number of MCMC samples to deliver downstream}

\item{random.seed}{Random number seed}

\item{nthread}{Number of CPU threads to employ}

\item{schedule}{Either \link{schedule_local} (execute locally), \link{schedule_pbs} or \link{schedule_slurm} (prepare for submission to HPC cluster)}
}
\description{
Define control parameters for the seaMass-Σ Bayesian model.
}
\section{Functions}{
\itemize{
\item \code{sigma_control}: Generator function
}}

