% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_control.R
\docType{class}
\name{theta_control-class}
\alias{theta_control-class}
\alias{theta_control}
\title{Control parameters for seaMass-theta}
\usage{
theta_control(
  keep = "group.quants",
  summarise = c("assays", "groups"),
  plot = c("assay.means", "group.standards", "group.quants", "group.quants.pca"),
  model = "theta",
  nwarmup = 256,
  thin = 4,
  random.seed = 0
)
}
\arguments{
\item{keep}{Outputs to keep MCMC samples for, \code{NULL}}

\item{summarise}{Outputs to write csv summaries for, \code{NULL} or a subset of
\code{c("assays", "groups")}
Note, you must summarise or keep \code{"group.quants"} if you want to run seaMass-delta!}

\item{plot}{Outputs to plot, \code{NULL} or a subset of
\code{c("assay.means", "group.quants")}}

\item{nwarmup}{Number of MCMC warmup iterations to run for each chain}

\item{thin}{MCMC thinning factor}

\item{random.seed}{Random number seed}
}
\description{
Define control parameters for the seaMass-theta Bayesian model.
}
\section{Functions}{
\itemize{
\item \code{theta_control}: Generator function
}}

